HerbData.jl Documentation

HerbData.IOExample — Type

struct IOExample <: Example

An input-output example. input is a Dict of {Symbol,Any} where the symbol represents a variable in a program. output can be anything.

HerbData.IOPExample — Type

struct IOPExample <: Example

An input-output example with an associated program. ex is an IOExample. program is a program of arbitrary form. Please note that this is a pure container, and thus does not guarantee any checks on the validity of the program.

HerbData.Problem — Type

struct Problem

Program synthesis problem defined with a vector of Examples

HerbData.read_IOPexamples — Method

read_IOPexamples(filepath::AbstractString)::Vector{Tuple{IOPExample}

Reads serialized IO + program examples from disk after type checking.

HerbData.read_IOexamples — Method

read_IOexamples(filepath::AbstractString)::Vector{IOExample}

Reads serialized IO examples from disk after type checking.

HerbData.readdata — Method

readdata(directory::AbstractString, lineparser::Function)::Vector{Problem}

Reads all files in the given directory and parses them line by line into an ExampleProblem using the given lineparser.

TODO: Turn this into an iterator that doesn't load all data into memory at initialization.

HerbData.readfile — Method

readfile(filepath::AbstractString, lineparser::Function)::Problem

Reads a file and parses every non-empty line using the line parser.

HerbData.write_IOPexamples — Method

write_IOPexamples(filepath::AbstractString, examples::Vector{Tuple{IOExample, Any}})

Writes IO examples and the corresponding programs to disk by serializing them into a file using HDF5 checking for and appending the .xiop.

HerbData.write_IOexamples — Method

write_IOexamples(filepath::AbstractString, examples::Vector{IOExample})

Writes IO examples to disk by serializing them into a file using HDF5 checking for and appending the .xio file ending.

Index

HerbConstraints.AbstractMatchNode

HerbConstraints.ComesAfter

HerbConstraints.ComesAfter

HerbConstraints.Forbidden

HerbConstraints.ForbiddenPath

HerbConstraints.GrammarContext

HerbConstraints.LocalConstraint

HerbConstraints.LocalForbidden

HerbConstraints.LocalOneOf

HerbConstraints.LocalOrdered

HerbConstraints.MatchFail

HerbConstraints.MatchNode

HerbConstraints.MatchVar

HerbConstraints.OneOf

HerbConstraints.Ordered

HerbConstraints.PropagatorConstraint

HerbConstraints.RequireOnLeft

HerbCore.AbstractRuleNode

HerbCore.Constraint

HerbCore.Grammar

HerbCore.Hole

HerbCore.HoleReference

HerbCore.RuleNode

HerbCore.RuleNode

HerbCore.RuleNode

HerbData.IOExample

HerbData.IOPExample

HerbData.Problem

HerbGrammar.ContextFreeGrammar

HerbGrammar.ContextSensitiveGrammar

HerbGrammar.NodeLoc

HerbGrammar.SymbolTable

HerbGrammar.SymbolTable

HerbSearch.ContextSensitivePriorityEnumerator

HerbSearch.ExpandFailureReason

HerbSearch.ExpressionIterator

HerbSearch.GeneticSearchIterator

HerbSearch.PriorityQueueItem

HerbSearch.PriorityQueueItem

HerbSearch.PropagateResult

HerbSearch.StochasticSearchEnumerator

Base.get

Base.insert!

Base.isless

Base.iterate

Base.iterate

Base.iterate

Base.iterate

Base.iterate

Base.length

Base.length

Base.rand

Base.rand

Base.rand

Base.show

Base.show

HerbConstraints._matchnode2expr

HerbConstraints._matchnode2expr

HerbConstraints._matchnode2expr

HerbConstraints._matchnode2expr

HerbConstraints._pattern_match

HerbConstraints._pattern_match

HerbConstraints._pattern_match_with_hole

HerbConstraints._pattern_match_with_hole

HerbConstraints._pattern_match_with_hole

HerbConstraints._pattern_match_with_hole

HerbConstraints._rulenode_compare

HerbConstraints._rulenode_match_with_hole

HerbConstraints.addparent!

HerbConstraints.annotation2constraint

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.copy_and_insert

HerbConstraints.make_smaller_or_equal

HerbConstraints.matchnode2expr

HerbConstraints.matchnode2expr

HerbConstraints.propagate

HerbConstraints.propagate

HerbConstraints.propagate

HerbConstraints.propagate

HerbConstraints.propagate

HerbConstraints.propagate

HerbConstraints.propagate

HerbConstraints.propagate

HerbCore.contains_hole

HerbCore.depth

HerbCore.get_node_at_location

HerbCore.get_rulesequence

HerbCore.node_depth

HerbCore.rulesoftype

HerbCore.rulesonleft

HerbCore.swap_node

HerbCore.swap_node

HerbData.read_IOPexamples

HerbData.read_IOexamples

HerbData.readdata

HerbData.readfile

HerbData.write_IOPexamples

HerbData.write_IOexamples

HerbGrammar.add_rule!

HerbGrammar.addconstraint!

HerbGrammar.cfg2csg

HerbGrammar.child_types

HerbGrammar.child_types

HerbGrammar.cleanup_removed_rules!

HerbGrammar.clearconstraints!

HerbGrammar.containedin

HerbGrammar.contains_returntype

HerbGrammar.expr2cfgrammar

HerbGrammar.expr2csgrammar

HerbGrammar.expr2pcfgrammar

HerbGrammar.expr2pcsgrammar

HerbGrammar.get_childtypes

HerbGrammar.get_domain

HerbGrammar.get_domain

HerbGrammar.get_node_at_location

HerbGrammar.get_rulesequence

HerbGrammar.has_children

HerbGrammar.iscomplete

HerbGrammar.iseval

HerbGrammar.iseval

HerbGrammar.iseval

HerbGrammar.isprobabilistic

HerbGrammar.isterminal

HerbGrammar.isterminal

HerbGrammar.isterminal

HerbGrammar.isvariable

HerbGrammar.log_probability

HerbGrammar.max_arity

HerbGrammar.mindepth

HerbGrammar.mindepth_map

HerbGrammar.nchildren

HerbGrammar.nchildren

HerbGrammar.nonterminals

HerbGrammar.probability

HerbGrammar.read_cfg

HerbGrammar.read_csg

HerbGrammar.read_pcfg

HerbGrammar.read_pcsg

HerbGrammar.remove_rule!

HerbGrammar.return_type

HerbGrammar.return_type

HerbGrammar.root_node_loc

HerbGrammar.rulenode2expr

HerbGrammar.rulenode_log_probability

HerbGrammar.rulesoftype

HerbGrammar.rulesoftype

HerbGrammar.rulesoftype

HerbGrammar.rulesonleft

HerbGrammar.store_cfg

HerbGrammar.store_csg

HerbGrammar.subsequenceof

HerbGrammar.swap_node

HerbGrammar.swap_node

HerbInterpret.evaluate_program

HerbInterpret.execute_on_examples

HerbInterpret.interpret

HerbInterpret.test_all_examples

HerbInterpret.test_examples

HerbInterpret.test_with_input

HerbSearch._expand

HerbSearch._expand

HerbSearch._find_next_complete_tree

HerbSearch.best_accept

HerbSearch.calculate_cost

HerbSearch.const_temperature

HerbSearch.constructNeighbourhood

HerbSearch.constructNeighbourhoodRuleSubset

HerbSearch.count_expressions

HerbSearch.count_expressions

HerbSearch.crossover_swap_children_1

HerbSearch.crossover_swap_children_2

HerbSearch.decreasing_temperature

HerbSearch.default_error_function

HerbSearch.default_fitness

HerbSearch.enumerate_neighbours_propose

HerbSearch.get_best_program

HerbSearch.get_bfs_enumerator

HerbSearch.get_bfs_enumerator

HerbSearch.get_dfs_enumerator

HerbSearch.get_dfs_enumerator

HerbSearch.get_genetic_enumerator

HerbSearch.get_mh_enumerator

HerbSearch.get_most_likely_first_enumerator

HerbSearch.get_most_likely_first_enumerator

HerbSearch.get_sa_enumerator

HerbSearch.get_vlsn_enumerator

HerbSearch.heuristic_leftmost

HerbSearch.heuristic_random

HerbSearch.heuristic_rightmost

HerbSearch.heuristic_smallest_domain

HerbSearch.mean_squared_error

HerbSearch.misclassification

HerbSearch.most_likely_priority_function

HerbSearch.mse_error_function

HerbSearch.mutate_random!

HerbSearch.probabilistic_accept

HerbSearch.probabilistic_accept_with_temperature

HerbSearch.probabilistic_accept_with_temperature_fraction

HerbSearch.propagate_constraints

HerbSearch.random_fill_propose

HerbSearch.search

HerbSearch.search_best

HerbSearch.search_rulenode

HerbSearch.select_chromosome

HerbSearch.select_fitness_proportional_parents

HerbSearch.validate_iterator

StatsBase.sample

StatsBase.sample

StatsBase.sample

StatsBase.sample

HerbConstraints.@csgrammar_annotated

HerbGrammar.@cfgrammar

HerbGrammar.@csgrammar

HerbGrammar.@pcfgrammar

HerbGrammar.@pcsgrammar