HerbGrammar.jl Documentation

HerbGrammar.ContextFreeGrammar — Type

ContextFreeGrammar <: Grammar

Represents a context-free grammar and its production rules. Consists of:

rules::Vector{Any}: A list of RHS of rules (subexpressions).

types::Vector{Symbol}: A list of LHS of rules (types, all symbols).

isterminal::BitVector: A bitvector where bit i represents whether rule i is terminal.

iseval::BitVector: A bitvector where bit i represents whether rule i is an eval rule.

bytype::Dict{Symbol,Vector{Int}}: A dictionary that maps a type to all rules of said type.

domains::Dict{Symbol, BitVector}: A dictionary that maps a type to a domain bitvector. The domain bitvector has bit i set to true iff the ith rule is of this type.

childtypes::Vector{Vector{Symbol}}: A list of types of the children for each rule. If a rule is terminal, the corresponding list is empty.

log_probabilities::Union{Vector{Real}, Nothing}: A list of probabilities for each rule. If the grammar is non-probabilistic, the list can be nothing.

Use the @cfgrammar macro to create a ContextFreeGrammar object. Use the @pcfgrammar macro to create a ContextFreeGrammar object with probabilities. For context-sensitive grammars, see ContextSensitiveGrammar.

HerbGrammar.ContextSensitiveGrammar — Type

ContextSensitiveGrammar <: Grammar

Represents a context-sensitive grammar. Extends Grammar with constraints.

Consists of:

rules::Vector{Any}: A list of RHS of rules (subexpressions).

types::Vector{Symbol}: A list of LHS of rules (types, all symbols).

isterminal::BitVector: A bitvector where bit i represents whether rule i is terminal.

iseval::BitVector: A bitvector where bit i represents whether rule i is an eval rule.

bytype::Dict{Symbol,Vector{Int}}: A dictionary that maps a type to all rules of said type.

domains::Dict{Symbol, BitVector}: A dictionary that maps a type to a domain bitvector. The domain bitvector has bit i set to true iff the ith rule is of this type.

childtypes::Vector{Vector{Symbol}}: A list of types of the children for each rule. If a rule is terminal, the corresponding list is empty.

log_probabilities::Union{Vector{Real}, Nothing}: A list of probabilities for each rule. If the grammar is non-probabilistic, the list can be nothing.

constraints::Vector{Constraint}: A list of constraints that programs in this grammar have to abide.

Use the @csgrammar macro to create a ContextSensitiveGrammar object. Use the @pcsgrammar macro to create a ContextSensitiveGrammar object with probabilities. For context-free grammars, see ContextFreeGrammar.

HerbGrammar.NodeLoc — Type

NodeLoc A helper struct that points to a node in the tree via its parent such that the child can be easily swapped out. If i is 0 the node pointed to is the root node and parent is the node itself.

HerbGrammar.SymbolTable — Type

SymbolTable(grammar::Grammar, mod::Module=Main)

Returns a SymbolTable populated with a mapping from symbols in the Grammar to symbols in module mod or Main, if defined.

HerbGrammar.SymbolTable — Type

SymbolTable

Data structure for mapping terminal symbols in the Grammar to their Julia interpretation.

HerbGrammar.@cfgrammar — Macro

@cfgrammar

A macro for defining a ContextFreeGrammar.

Example usage:

grammar = @cfgrammar begin
	R = x
	R = 1 | 2
	R = R + R
end

Syntax:

Literals: Symbols that are already defined in Julia are considered literals, such as 1, 2, or π. For example: R = 1.

Variables: A variable is a symbol that is not a nonterminal symbol and not already defined in Julia. For example: R = x.

Functions: Functions and infix operators that are defined in Julia or the Main module can be used with the default evaluator. For example: R = R + R, R = f(a, b).

Combinations: Multiple rules can be defined on a single line in the grammar definition using the | symbol. For example: R = 1 | 2 | 3.

Iterators: Another way to define multiple rules is by providing a Julia iterator after a | symbol. For example: R = |(1:9).

Related:

@csgrammar uses the same syntax to create ContextSensitiveGrammars.

@pcfgrammar uses a similar syntax to create probabilistic ContextFreeGrammars.

HerbGrammar.@csgrammar — Macro

@csgrammar

A macro for defining a ContextSensitiveGrammar. Constraints can be added afterwards using the addconstraint! function.

Example usage:

grammar = @csgrammar begin
	R = x
	R = 1 | 2
	R = R + R
end

Syntax:

Literals: Symbols that are already defined in Julia are considered literals, such as 1, 2, or π. For example: R = 1.

Variables: A variable is a symbol that is not a nonterminal symbol and not already defined in Julia. For example: R = x.

Functions: Functions and infix operators that are defined in Julia or the Main module can be used with the default evaluator. For example: R = R + R, R = f(a, b).

Combinations: Multiple rules can be defined on a single line in the grammar definition using the | symbol. For example: R = 1 | 2 | 3.

Iterators: Another way to define multiple rules is by providing a Julia iterator after a | symbol. For example: R = |(1:9).

Related:

@cfgrammar uses the same syntax to create ContextFreeGrammars.

@pcsgrammar uses a similar syntax to create probabilistic ContextSensitiveGrammars.

HerbGrammar.@pcfgrammar — Macro

@pcfgrammar

A macro for defining a probabilistic ContextFreeGrammar.

Example usage:

grammar = @pcfgrammar begin
	0.5 : R = x
	0.3 : R = 1 | 2
	0.2 : R = R + R
end

Syntax:

The syntax of rules is identical to the syntax used by @cfgrammar:

Literals: Symbols that are already defined in Julia are considered literals, such as 1, 2, or π. For example: R = 1.

Variables: A variable is a symbol that is not a nonterminal symbol and not already defined in Julia. For example: R = x.

Functions: Functions and infix operators that are defined in Julia or the Main module can be used with the default evaluator. For example: R = R + R, R = f(a, b).

Combinations: Multiple rules can be defined on a single line in the grammar definition using the | symbol. For example: R = 1 | 2 | 3.

Iterators: Another way to define multiple rules is by providing a Julia iterator after a | symbol. For example: R = |(1:9).

Every rule is also prefixed with a probability. Rules and probabilities are separated using the : symbol. If multiple rules are defined on a single line, the probability is equally divided between the rules. The sum of probabilities for all rules of a certain non-terminal symbol should be equal to 1. The probabilities are automatically scaled if this isn't the case.

Related:

@pcsgrammar uses the same syntax to create probabilistic ContextSensitiveGrammars.

@cfgrammar uses a similar syntax to create non-probabilistic ContextFreeGrammars.

HerbGrammar.@pcsgrammar — Macro

@pcsgrammar

A macro for defining a probabilistic ContextSensitiveGrammar.

Example usage:

grammar = @pcsgrammar begin
	0.5 : R = x
	0.3 : R = 1 | 2
	0.2 : R = R + R
end

Syntax:

The syntax of rules is identical to the syntax used by @csgrammar:

Literals: Symbols that are already defined in Julia are considered literals, such as 1, 2, or π. For example: R = 1.

Variables: A variable is a symbol that is not a nonterminal symbol and not already defined in Julia. For example: R = x.

Functions: Functions and infix operators that are defined in Julia or the Main module can be used with the default evaluator. For example: R = R + R, R = f(a, b).

Combinations: Multiple rules can be defined on a single line in the grammar definition using the | symbol. For example: R = 1 | 2 | 3.

Iterators: Another way to define multiple rules is by providing a Julia iterator after a | symbol. For example: R = |(1:9).

Every rule is also prefixed with a probability. Rules and probabilities are separated using the : symbol. If multiple rules are defined on a single line, the probability is equally divided between the rules. The sum of probabilities for all rules of a certain non-terminal symbol should be equal to 1. The probabilities are automatically scaled if this isn't the case.

Related:

@pcfgrammar uses the same syntax to create probabilistic ContextFreeGrammars.

@csgrammar uses a similar syntax to create non-probabilistic ContextSensitiveGrammars.

Base.get — Method

get(root::RuleNode, loc::NodeLoc) Obtain the node pointed to by loc.

Base.insert! — Method

insert!(loc::NodeLoc, rulenode::RuleNode) Replaces the subtree pointed to by loc with the given rulenode.

HerbGrammar.add_rule! — Method

add_rule!(g::Grammar, e::Expr)

Adds a rule to the grammar.

Usage:

add_rule!(grammar, :("Real = Real + Real"))

The syntax is identical to the syntax of @csgrammar and @cfgrammar, but only single rules are supported.

Warning

Calls to this function are ignored if a rule is already in the grammar.

HerbGrammar.addconstraint! — Method

addconstraint!(grammar::ContextSensitiveGrammar, c::Constraint)

Adds a Constraint to a ContextSensitiveGrammar.

HerbGrammar.cfg2csg — Method

cfg2csg(g::ContextFreeGrammar)::ContextSensitiveGrammar

Converts a ContextFreeGrammar to a ContextSensitiveGrammar without any Constraints.

HerbGrammar.child_types — Method

child_types(grammar::Grammar, rule_index::Int)

Returns the types of the children (nonterminals) of the production rule at rule_index.

HerbGrammar.child_types — Method

child_types(grammar::Grammar, node::RuleNode)

Returns the list of child types (nonterminal symbols) in the production rule used by node.

HerbGrammar.cleanup_removed_rules! — Method

cleanup_removed_rules!(g::Grammar)

Removes any placeholders for previously deleted rules. This means that indices get shifted.

Warning

When indices are shifted, this grammar can no longer be used to interpret AbstractRuleNode trees created before the call to this function. These trees become meaningless.

HerbGrammar.clearconstraints! — Method

Clear all constraints from the grammar

HerbGrammar.containedin — Method

containedin(vec1::Vector, vec2::Vector)

Checks if elements of vec1 are contained in vec2 in the same order (possibly with elements in between)

HerbGrammar.contains_returntype — Function

contains_returntype(node::RuleNode, grammar::Grammar, sym::Symbol, maxdepth::Int=typemax(Int))

Returns true if the tree rooted at node contains at least one node at depth less than maxdepth with the given return type or nonterminal symbol.

HerbGrammar.expr2cfgrammar — Method

expr2cfgrammar(ex::Expr)::ContextFreeGrammar

A function for converting an Expr to a ContextFreeGrammar. If the expression is hardcoded, you should use the @cfgrammar macro. Only expressions in the correct format (see @cfgrammar) can be converted.

Example usage:

grammar = expr2cfgrammar(
	begin
		R = x
		R = 1 | 2
		R = R + R
	end
)

HerbGrammar.expr2csgrammar — Method

expr2csgrammar(ex::Expr)::ContextSensitiveGrammar

A function for converting an Expr to a ContextSensitiveGrammar. If the expression is hardcoded, you should use the @csgrammar macro. Only expressions in the correct format (see @csgrammar) can be converted.

Example usage:

grammar = expr2csgrammar(
	begin
		R = x
		R = 1 | 2
		R = R + R
	end
)

HerbGrammar.expr2pcfgrammar — Method

Function for converting an Expr to a ContextFreeGrammar with probabilities. If the expression is hardcoded, you should use the @pcfgrammar macro. Only expressions in the correct format (see @pcfgrammar) can be converted.

Example usage:

grammar = expr2pcsgrammar(
	begin
		0.5 : R = x
		0.3 : R = 1 | 2
		0.2 : R = R + R
	end
)

HerbGrammar.expr2pcsgrammar — Method

Function for converting an Expr to a ContextSensitiveGrammar with probabilities. If the expression is hardcoded, you should use the @pcsgrammar macro. Only expressions in the correct format (see @pcsgrammar) can be converted.

Example usage:

grammar = expr2pcsgrammar(
	begin
		0.5 : R = x
		0.3 : R = 1 | 2
		0.2 : R = R + R
	end
)

HerbGrammar.get_childtypes — Method

get_childtypes(rule::Any, types::AbstractVector{Symbol})

Returns the child types/nonterminals of a production rule.

HerbGrammar.get_domain — Method

get_domain(g::Grammar, type::Symbol)::BitVector

Returns the domain for the hole of a certain type as a BitVector of the same length as the number of rules in the grammar. Bit i is set to true iff rule i is of type type.

Info

Since this function can be intensively used when exploring a program space defined by a grammar, the outcomes of this function are precomputed and stored in the domains field in a Grammar.

HerbGrammar.get_domain — Method

get_domain(g::Grammar, rules::Vector{Int})::BitVector

Takes a domain rules defined as a vector of ints and converts it to a domain defined as a BitVector.

HerbGrammar.get_node_at_location — Method

get_node_at_location(root::RuleNode, location::Vector{Int})

Retrieves a RuleNode at the given location by reference.

HerbGrammar.get_rulesequence — Method

get_rulesequence(node::RuleNode, path::Vector{Int})

Extract the derivation sequence from a path (sequence of child indices) and an AbstractRuleNode. If the path is deeper than the deepest node, it returns what it has.

HerbGrammar.has_children — Method

has_children(node::RuleNode)

Returns true if node has children

HerbGrammar.iscomplete — Method

iscomplete(grammar::Grammar, node::RuleNode)

Returns true if the expression represented by the RuleNode is a complete expression, meaning that it is fully defined and doesn't have any Holes.

HerbGrammar.iseval — Method

iseval(rule)

Returns true if the rule is the special evaluate immediately function, i.e., _()

Compat

evaluate immediately functionality is not yet supported by most of Herb.jl

HerbGrammar.iseval — Method

iseval(grammar::Grammar, index::Int)::Bool

Returns true if the production rule at rule_index contains the special _() eval function.

Compat

evaluate immediately functionality is not yet supported by most of Herb.jl

HerbGrammar.iseval — Method

iseval(grammar::Grammar)::Bool

Returns true if any production rules in grammar contain the special _() eval function.

Compat

evaluate immediately functionality is not yet supported by most of Herb.jl

HerbGrammar.isprobabilistic — Method

isprobabilistic(grammar::Grammar)::Bool

Function returns whether a Grammar is probabilistic.

HerbGrammar.isterminal — Method

isterminal(rule::Any, types::AbstractVector{Symbol})

Returns true if the rule is terminal, i.e., it does not contain any of the types in the provided vector. For example, :(x) is terminal, and :(1+1) is terminal, but :(Real + Real) is typically not.

HerbGrammar.isterminal — Method

isterminal(grammar::Grammar, rule_index::Int)::Bool

Returns true if the production rule at rule_index is terminal, i.e., does not contain any nonterminal symbols.

HerbGrammar.isterminal — Method

isterminal(grammar::Grammar, node::RuleNode)::Bool

Returns true if the production rule used by node is terminal, i.e., does not contain any nonterminal symbols.

HerbGrammar.isvariable — Method

isvariable(grammar::Grammar, node::RuleNode)::Bool

Returns true if the rule used by node represents a variable.

HerbGrammar.log_probability — Method

log_probability(grammar::Grammar, index::Int)::Real

Returns the log probability for the rule at index in the grammar.

Warning

If the grammar is not probabilistic, a warning is displayed, and a uniform probability is assumed.

HerbGrammar.max_arity — Method

max_arity(grammar::Grammar)::Int

Returns the maximum arity (number of children) over all production rules in the Grammar.

HerbGrammar.mindepth — Method

mindepth(grammar::Grammar, typ::Symbol, dmap::AbstractVector{Int})

Returns the minimum depth achievable for a given nonterminal symbol. The minimum depth is the depth of the lowest tree that can be made using typ as a start symbol. dmap can be obtained from mindepth_map.

HerbGrammar.mindepth_map — Method

mindepth_map(grammar::Grammar)

Returns the minimum depth achievable for each production rule in the Grammar. In other words, this function finds the depths of the lowest trees that can be made using each of the available production rules as a root.

HerbGrammar.nchildren — Method

nchildren(grammar::Grammar, rule_index::Int)::Int

Returns the number of children (nonterminals) of the production rule at rule_index.

HerbGrammar.nchildren — Method

nchildren(grammar::Grammar, node::RuleNode)::Int

Returns the number of children in the production rule used by node.

HerbGrammar.nonterminals — Method

nonterminals(grammar::Grammar)::Vector{Symbol}

Returns a list of the nonterminals or types in the Grammar.

HerbGrammar.probability — Method

probability(grammar::Grammar, index::Int)::Real

Return the probability for a rule in the grammar. Use log_probability whenever possible.

Warning

If the grammar is not probabilistic, a warning is displayed, and a uniform probability is assumed.

HerbGrammar.read_cfg — Method

read_cfg(filepath::AbstractString)::ContextFreeGrammar

Reads a ContextFreeGrammar from the file provided in filepath.

Danger

Only open trusted grammars. Parts of the grammar can be passed to Julia's eval function.

HerbGrammar.read_csg — Method

read_csg(grammarpath::AbstractString, constraintspath::AbstractString)::ContextSensitiveGrammar

Reads a ContextSensitiveGrammar from the files at grammarpath and constraintspath. The grammar path may also point to a ContextFreeGrammar.

Danger

Only open trusted grammars. Parts of the grammar can be passed to Julia's eval function.

HerbGrammar.read_pcfg — Method

read_pcfg(filepath::AbstractString)::ContextFreeGrammar

Reads a probabilistic ContextFreeGrammar from a file provided in filepath.

Danger

Only open trusted grammars. Parts of the grammar can be passed to Julia's eval function.

HerbGrammar.read_pcsg — Method

read_pcsg(grammarpath::AbstractString, constraintspath::AbstractString)::ContextSensitiveGrammar

Reads a probabilistic ContextSensitiveGrammar from the files at grammarpath and constraintspath. The grammar path may also point to a ContextFreeGrammar.

Danger

Only open trusted grammars. Parts of the grammar can be passed to Julia's eval function.

HerbGrammar.remove_rule! — Method

remove_rule!(g::Grammar, idx::Int)

Removes the rule corresponding to idx from the grammar. In order to avoid shifting indices, the rule is replaced with nothing, and all other data structures are updated accordingly.

HerbGrammar.return_type — Method

return_type(grammar::Grammar, rule_index::Int)::Symbol

Returns the type of the production rule at rule_index.

HerbGrammar.return_type — Method

return_type(grammar::Grammar, node::RuleNode)

Gives the return type or nonterminal symbol in the production rule used by node.

HerbGrammar.root_node_loc — Method

rootnodeloc(root::RuleNode) Returns a NodeLoc pointing to the root node.

HerbGrammar.rulenode2expr — Method

rulenode2expr(rulenode::RuleNode, grammar::Grammar)

Converts a RuleNode into a Julia expression corresponding to the rule definitions in the grammar. The returned expression can be evaluated with Julia semantics using eval().

HerbGrammar.rulenode_log_probability — Method

Calculates the log probability associated with a rulenode in a probabilistic grammar.

HerbGrammar.rulesoftype — Method

rulesoftype(node::RuleNode, grammar::Grammar, ruletype::Symbol, ignoreNode::RuleNode)

Returns every rule of nonterminal symbol ruletype that is also used in the AbstractRuleNode tree, but not in the ignoreNode subtree.

Warning

The ignoreNode must be a subtree of node for it to have an effect.

HerbGrammar.rulesoftype — Method

rulesoftype(node::RuleNode, grammar::Grammar, ruletype::Symbol)

Returns every rule of nonterminal symbol ruletype that is also used in the AbstractRuleNode tree.

HerbGrammar.rulesoftype — Method

rulesoftype(node::RuleNode, ruleset::Set{Int}, ignoreNode::RuleNode)

Returns every rule in the ruleset that is also used in the AbstractRuleNode tree, but not in the ignoreNode subtree.

Warning

The ignoreNode must be a subtree of node for it to have an effect.

HerbGrammar.rulesonleft — Method

rulesonleft(expr::RuleNode, path::Vector{Int})::Set{Int}

Finds all rules that are used in the left subtree defined by the path.

HerbGrammar.store_cfg — Method

store_cfg(filepath::AbstractString, grammar::ContextFreeGrammar)

Writes a ContextFreeGrammar to the file provided by filepath.

HerbGrammar.store_csg — Method

store_csg(grammarpath::AbstractString, constraintspath::AbstractString, g::ContextSensitiveGrammar)

Writes a ContextSensitiveGrammar to the files at grammarpath and constraintspath. The grammarpath file will contain a ContextSensitiveGrammar definition, and the constraintspath file will contain the Constraints of the ContextSensitiveGrammar.

HerbGrammar.subsequenceof — Method

subsequenceof(vec1::Vector{Int}, vec2::Vector{Int})

Checks if vec1 is a subsequence of vec2.

HerbGrammar.swap_node — Method

swap_node(expr::AbstractRuleNode, new_expr::AbstractRuleNode, path::Vector{Int})

Replace a node in expr, specified by path, with new_expr. Path is a sequence of child indices, starting from the root node.

HerbGrammar.swap_node — Method

swap_node(expr::RuleNode, node::RuleNode, child_index::Int, new_expr::RuleNode)

Replace child i of a node, a part of larger expr, with new_expr.

Index

HerbConstraints.AbstractMatchNode

HerbConstraints.ComesAfter

HerbConstraints.ComesAfter

HerbConstraints.Forbidden

HerbConstraints.ForbiddenPath

HerbConstraints.GrammarContext

HerbConstraints.LocalConstraint

HerbConstraints.LocalForbidden

HerbConstraints.LocalOneOf

HerbConstraints.LocalOrdered

HerbConstraints.MatchFail

HerbConstraints.MatchNode

HerbConstraints.MatchVar

HerbConstraints.OneOf

HerbConstraints.Ordered

HerbConstraints.PropagatorConstraint

HerbConstraints.RequireOnLeft

HerbCore.AbstractRuleNode

HerbCore.Constraint

HerbCore.Grammar

HerbCore.Hole

HerbCore.HoleReference

HerbCore.RuleNode

HerbCore.RuleNode

HerbCore.RuleNode

HerbData.IOExample

HerbData.IOPExample

HerbData.Problem

HerbGrammar.ContextFreeGrammar

HerbGrammar.ContextSensitiveGrammar

HerbGrammar.NodeLoc

HerbGrammar.SymbolTable

HerbGrammar.SymbolTable

HerbSearch.ContextSensitivePriorityEnumerator

HerbSearch.ExpandFailureReason

HerbSearch.ExpressionIterator

HerbSearch.GeneticSearchIterator

HerbSearch.PriorityQueueItem

HerbSearch.PriorityQueueItem

HerbSearch.PropagateResult

HerbSearch.StochasticSearchEnumerator

Base.get

Base.insert!

Base.isless

Base.iterate

Base.iterate

Base.iterate

Base.iterate

Base.iterate

Base.length

Base.length

Base.rand

Base.rand

Base.rand

Base.show

Base.show

HerbConstraints._matchnode2expr

HerbConstraints._matchnode2expr

HerbConstraints._matchnode2expr

HerbConstraints._matchnode2expr

HerbConstraints._pattern_match

HerbConstraints._pattern_match

HerbConstraints._pattern_match_with_hole

HerbConstraints._pattern_match_with_hole

HerbConstraints._pattern_match_with_hole

HerbConstraints._pattern_match_with_hole

HerbConstraints._rulenode_compare

HerbConstraints._rulenode_match_with_hole

HerbConstraints.addparent!

HerbConstraints.annotation2constraint

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.copy_and_insert

HerbConstraints.make_smaller_or_equal

HerbConstraints.matchnode2expr

HerbConstraints.matchnode2expr

HerbConstraints.propagate

HerbConstraints.propagate

HerbConstraints.propagate

HerbConstraints.propagate

HerbConstraints.propagate

HerbConstraints.propagate

HerbConstraints.propagate

HerbConstraints.propagate

HerbCore.contains_hole

HerbCore.depth

HerbCore.get_node_at_location

HerbCore.get_rulesequence

HerbCore.node_depth

HerbCore.rulesoftype

HerbCore.rulesonleft

HerbCore.swap_node

HerbCore.swap_node

HerbData.read_IOPexamples

HerbData.read_IOexamples

HerbData.readdata

HerbData.readfile

HerbData.write_IOPexamples

HerbData.write_IOexamples

HerbGrammar.add_rule!

HerbGrammar.addconstraint!

HerbGrammar.cfg2csg

HerbGrammar.child_types

HerbGrammar.child_types

HerbGrammar.cleanup_removed_rules!

HerbGrammar.clearconstraints!

HerbGrammar.containedin

HerbGrammar.contains_returntype

HerbGrammar.expr2cfgrammar

HerbGrammar.expr2csgrammar

HerbGrammar.expr2pcfgrammar

HerbGrammar.expr2pcsgrammar

HerbGrammar.get_childtypes

HerbGrammar.get_domain

HerbGrammar.get_domain

HerbGrammar.get_node_at_location

HerbGrammar.get_rulesequence

HerbGrammar.has_children

HerbGrammar.iscomplete

HerbGrammar.iseval

HerbGrammar.iseval

HerbGrammar.iseval

HerbGrammar.isprobabilistic

HerbGrammar.isterminal

HerbGrammar.isterminal

HerbGrammar.isterminal

HerbGrammar.isvariable

HerbGrammar.log_probability

HerbGrammar.max_arity

HerbGrammar.mindepth

HerbGrammar.mindepth_map

HerbGrammar.nchildren

HerbGrammar.nchildren

HerbGrammar.nonterminals

HerbGrammar.probability

HerbGrammar.read_cfg

HerbGrammar.read_csg

HerbGrammar.read_pcfg

HerbGrammar.read_pcsg

HerbGrammar.remove_rule!

HerbGrammar.return_type

HerbGrammar.return_type

HerbGrammar.root_node_loc

HerbGrammar.rulenode2expr

HerbGrammar.rulenode_log_probability

HerbGrammar.rulesoftype

HerbGrammar.rulesoftype

HerbGrammar.rulesoftype

HerbGrammar.rulesonleft

HerbGrammar.store_cfg

HerbGrammar.store_csg

HerbGrammar.subsequenceof

HerbGrammar.swap_node

HerbGrammar.swap_node

HerbInterpret.evaluate_program

HerbInterpret.execute_on_examples

HerbInterpret.interpret

HerbInterpret.test_all_examples

HerbInterpret.test_examples

HerbInterpret.test_with_input

HerbSearch._expand

HerbSearch._expand

HerbSearch._find_next_complete_tree

HerbSearch.best_accept

HerbSearch.calculate_cost

HerbSearch.const_temperature

HerbSearch.constructNeighbourhood

HerbSearch.constructNeighbourhoodRuleSubset

HerbSearch.count_expressions

HerbSearch.count_expressions

HerbSearch.crossover_swap_children_1

HerbSearch.crossover_swap_children_2

HerbSearch.decreasing_temperature

HerbSearch.default_error_function

HerbSearch.default_fitness

HerbSearch.enumerate_neighbours_propose

HerbSearch.get_best_program

HerbSearch.get_bfs_enumerator

HerbSearch.get_bfs_enumerator

HerbSearch.get_dfs_enumerator

HerbSearch.get_dfs_enumerator

HerbSearch.get_genetic_enumerator

HerbSearch.get_mh_enumerator

HerbSearch.get_most_likely_first_enumerator

HerbSearch.get_most_likely_first_enumerator

HerbSearch.get_sa_enumerator

HerbSearch.get_vlsn_enumerator

HerbSearch.heuristic_leftmost

HerbSearch.heuristic_random

HerbSearch.heuristic_rightmost

HerbSearch.heuristic_smallest_domain

HerbSearch.mean_squared_error

HerbSearch.misclassification

HerbSearch.most_likely_priority_function

HerbSearch.mse_error_function

HerbSearch.mutate_random!

HerbSearch.probabilistic_accept

HerbSearch.probabilistic_accept_with_temperature

HerbSearch.probabilistic_accept_with_temperature_fraction

HerbSearch.propagate_constraints

HerbSearch.random_fill_propose

HerbSearch.search

HerbSearch.search_best

HerbSearch.search_rulenode

HerbSearch.select_chromosome

HerbSearch.select_fitness_proportional_parents

HerbSearch.validate_iterator

StatsBase.sample

StatsBase.sample

StatsBase.sample

StatsBase.sample

HerbConstraints.@csgrammar_annotated

HerbGrammar.@cfgrammar

HerbGrammar.@csgrammar

HerbGrammar.@pcfgrammar

HerbGrammar.@pcsgrammar