HerbSearch.jl Documentation

HerbSearch.BFSIterator — Type

@programiterator BFSIterator() <: TopDownIterator

Returns a breadth-first iterator given a grammar and a starting symbol. Returns trees in the grammar in increasing order of size. Inherits all stop-criteria from TopDownIterator.

HerbSearch.DFSIterator — Type

@programiterator DFSIterator() <: TopDownIterator

Returns a depth-first search enumerator given a grammar and a starting symbol. Returns trees in the grammar in decreasing order of size. Inherits all stop-criteria from TopDownIterator.

HerbSearch.ExpandFailureReason — Type

@enum ExpandFailureReason limit_reached=1 already_complete=2

Representation of the different reasons why expanding a partial tree failed. Currently, there are two possible causes of the expansion failing:

limit_reached: The depth limit or the size limit of the partial tree would be violated by the expansion

already_complete: There is no hole left in the tree, so nothing can be expanded.

HerbSearch.GeneticSearchIterator — Type

GeneticSearchIterator{FitnessFunction,CrossOverFunction,MutationFunction,SelectParentsFunction,EvaluationFunction} <: ProgramIterator

Defines an ProgramIterator using genetic search.

Consists of:

examples::Vector{<:IOExample}: a collection of examples defining the specification

evaluation_function::EvaluationFunction: interpreter to evaluate the individual programs

population_size::Int64: number of inviduals in the population

mutation_probability::Float64: probability of mutation for each individual

maximum_initial_population_depth::Int64: maximum depth of trees when population is initialized

end

HerbSearch.MHSearchIterator — Type

MHSearchIterator(examples::AbstractArray{<:IOExample}, cost_function::Function, evaluation_function::Function=HerbInterpret.execute_on_input)

Returns an enumerator that runs according to the Metropolis Hastings algorithm.

spec : array of examples

cost_function : cost function to evaluate the programs proposed

evaluation_function : evaluation function that evaluates the program generated and produces an output

The propose function is randomfillpropose and the accept function is probabilistic. The temperature value of the algorithm remains constant over time.

HerbSearch.MLFSIterator — Type

@programiterator MLFSIterator() <: TopDownIterator

Iterator that enumerates expressions in the grammar in decreasing order of probability (Only use this iterator with probabilistic grammars). Inherits all stop-criteria from TopDownIterator.

HerbSearch.PriorityQueueItem — Type

struct PriorityQueueItem

Represents an item in the priority enumerator priority queue. An item contains of:

tree: A partial AST

size: The size of the tree. This is a cached value which prevents having to traverse the entire tree each time the size is needed.

constraints: The local constraints that apply to this tree. These constraints are enforced each time the tree is modified.

HerbSearch.PriorityQueueItem — Method

PriorityQueueItem(tree::AbstractRuleNode, size::Int)

Constructs PriorityQueueItem given only a tree and the size, but no constraints.

HerbSearch.ProgramIterator — Type

abstract type ProgramIterator

Generic iterator for all possible search strategies. All iterators are expected to have the following fields:

grammar::ContextSensitiveGrammar: the grammar to search over

sym::Symbol: defines the start symbol from which the search should be started

max_depth::Int: maximum depth of program trees

max_size::Int: maximum number of AbstractRuleNodes of program trees

max_time::Int: maximum time the iterator may take

max_enumerations::Int: maximum number of enumerations

HerbSearch.PropagateResult — Type

@enum PropagateResult tree_complete=1 tree_incomplete=2 tree_infeasible=3

Representation of the possible results of a constraint propagation. At the moment there are three possible outcomes:

tree_complete: The propagation was applied successfully and the tree does not contain any holes anymore. Thus no constraints can be applied anymore.

tree_incomplete: The propagation was applied successfully and the tree does contain more holes. Thus more constraints may be applied to further prune the respective domains.

tree_infeasible: The propagation was succesful, but there are holes with empty domains. Hence, the tree is now infeasible.

HerbSearch.SASearchIterator — Type

SASearchIterator(spec, cost_function, initial_temperature=1, temperature_decreasing_factor = 0.99, evaluation_function::Function=HerbInterpret.execute_on_input)

Returns an enumerator that runs according to the Simulated Annealing Search algorithm.

spec : array of examples

cost_function : cost function to evaluate the programs proposed

initial_temperature : the starting temperature of the algorithm

temperature_decreasing_factor : the decreasing factor of the temperature of the time

evaluation_function : evaluation function that evaluates the program generated and produces an output

The propose function is random_fill_propose (the same as for Metropolis Hastings). The accept function is probabilistic but takes into account the tempeerature too.

HerbSearch.StochasticSearchIterator — Type

abstract type StochasticSearchIterator <: ProgramIterator

A unified abstract type for the algorithms Metropolis Hastings, Very Large Scale Neighbourhood and Simulated Annealing. Each algorithm implements neighbourhood, propose, accept and temperature functions. Below the signatures of each function is shown.

Signatures

Returns a node location from the program that is the neighbourhood. It can also return other information using dict

neighbourhood(iter::T, current_program::RuleNode) where T <: StochasticSearchIterator -> (loc::NodeLocation, dict::Dict)

Proposes a list of programs using the location provided by neighbourhood and the dict.

propose(iter::T, current_program::RuleNode, neighbourhood_node_loc::NodeLoc, dmap::AbstractVector{Int}, dict::Union{Nothing,Dict{String,Any}}) where T <: StochasticSearchIterator -> Iter[RuleNode]

Based on the current program and possible cost and temperature it accepts the program or not. Usually we would always want to accept better programs but we might get stuck if we do so. That is why some implementations of the accept function accept with a probability costs that are worse. cost means how different are the outcomes of the program compared to the correct outcomes. The lower the cost the better the program performs on the examples. The cost is provided by the cost_function

accept(::T, currentcost::Real, nextcost::Real, temperature::Real) where T <: StochasticSearchIterator -> Bool

Returns the new temperature based on the previous temperature. Higher the temperature means that the algorithm will explore more.

temperature(::T, current_temperature::Real) where T <: StochasticSearchIterator -> Real

Returns the cost of the current program. It receives a list of tuples (expected, found) and gives back a cost.

cost_function(outcomes::Tuple{<:Number,<:Number}[]) -> Real

Fields

grammar::ContextSensitiveGrammar grammar that the algorithm uses

sym::Symbol the start symbol of the algorithm :Real or :Int

examples::Vector{IOExample} example used to check the program

cost_function::Function

initial_temperature::Real = 1

evaluation_function::Function that evaluates the julia expressions

An iterator over all possible expressions of a grammar up to maxdepth with start symbol sym. Also inherits all stop criteria like `maxdepthfromProgramIterator`.

HerbSearch.SynthResult — Type

@enum SynthResult optimal_program=1 suboptimal_program=2

Representation of the possible results of the synth procedure. At the moment there are two possible outcomes:

optimal_program: The synthesized program satisfies the entire program specification.

suboptimal_program: The synthesized program does not satisfy the entire program specification, but got the best score from the evaluator.

HerbSearch.TopDownIterator — Type

mutable struct TopDownIterator <: ProgramIterator

Enumerates a context-free grammar starting at Symbol sym with respect to the grammar up to a given depth and a given size. The exploration is done using the given priority function for derivations, and the expand function for discovered nodes. Concrete iterators may overload the following methods:

priority_function

derivation_heuristic

hole_heuristic

HerbSearch.VLSNSearchIterator — Type

VLSNSearchIterator(spec, cost_function, enumeration_depth = 2, evaluation_function::Function=HerbInterpret.execute_on_input) = StochasticSearchIterator(

Returns an iterator that runs according to the Very Large Scale Neighbourhood Search algorithm.

spec : array of examples

cost_function : cost function to evaluate the programs proposed

vlsn_neighbourhood_depth : the enumeration depth to search for a best program at a time

evaluation_function : evaluation function that evaluates the program generated and produces an output

The propose function consists of all possible programs of the given enumeration_depth. The accept function accepts the program with the lowest cost according to the cost_function. The temperature value of the algorithm remains constant over time.

HerbSearch.@programiterator — Macro

@programiterator

Canonical way of creating a program iterator. The macro automatically declares the expected fields listed in the ProgramIterator documentation. Syntax accepted by the macro is as follows (anything enclosed in square brackets is optional): @programiterator [mutable] <IteratorName>( <arg₁>, ..., <argₙ> ) [<: <SupertypeIterator>] Note that the macro emits an assertion that the SupertypeIterator is a subtype of ProgramIterator which otherwise throws an ArgumentError. If no supertype is given, the new iterator extends ProgramIterator directly. Each <argᵢ> may be (almost) any expression valid in a struct declaration, and they must be comma separated. One known exception is that an inner constructor must always be given using the extended function <name>(...) ... end syntax. The mutable keyword determines whether the declared struct is mutable.

Base.iterate — Method

Base.iterate(iter::GeneticSearchIterator, current_state::GeneticIteratorState)

Iterates the search space using a genetic algorithm. Takes the iterator and the current state to mutate and crossover random inviduals. Returns the best program-so-far and the state of the iterator.

Base.iterate — Method

Base.iterate(iter::GeneticSearchIterator)

Iterates the search space using a genetic algorithm. First generates a population sampling random programs. Returns the best program-so-far, and the state of the iterator.

Base.iterate — Method

Base.iterate(iter::StochasticSearchIterator, current_state::IteratorState)

The algorithm that constructs the iterator of StochasticSearchIterator. It has the following structure:

get a random node location -> location,dict = neighbourhood(current_program)

call propose on the current program getting a list of possbile replacements in the node location

iterate through all the possible replacements and perform the replacement in the current program

accept the new program by modifying the next_program or reject the new program

return the new next_program

Base.iterate — Method

Base.iterate(iter::TopDownIterator, pq::DataStructures.PriorityQueue)

Describes the iteration for a given TopDownIterator and a PriorityQueue over the grammar without enqueueing new items to the priority queue. Recursively returns the result for the priority queue.

Base.iterate — Method

Base.iterate(iter::TopDownIterator)

Describes the iteration for a given TopDownIterator over the grammar. The iteration constructs a PriorityQueue first and then prunes it propagating the active constraints. Recursively returns the result for the priority queue.

Base.rand — Function

rand(::Type{RuleNode}, grammar::AbstractGrammar, max_depth::Int=10)

Generates a random RuleNode of arbitrary type and maximum depth max_depth.

Base.rand — Function

rand(::Type{RuleNode}, grammar::AbstractGrammar, typ::Symbol, max_depth::Int=10)

Generates a random RuleNode of return type typ and maximum depth max_depth.

Base.rand — Function

rand(::Type{RuleNode}, grammar::AbstractGrammar, typ::Symbol, dmap::AbstractVector{Int}, max_depth::Int=10)

Generates a random RuleNode, i.e. an expression tree, of root type typ and maximum depth max_depth guided by a depth map dmap if possible.

HerbSearch._calculate_cost — Method

_calculate_cost(program::RuleNode, cost_function::Function, spec::AbstractVector{IOExample}, grammar::AbstractGrammar, evaluation_function::Function)

Returns the cost of the program using the examples and the cost_function. It first convert the program to an expression and evaluates it on all the examples.

HerbSearch._expand — Method

_expand(node::Hole, grammar::ContextSensitiveGrammar, ::Int, max_holes::Int, context::GrammarContext, iter::TopDownIterator)::Union{ExpandFailureReason, Vector{TreeConstraints}}

Expands a given hole that was found in _expand using the given derivation heuristic. Returns the list of discovered nodes in that order and with their respective constraints.

HerbSearch._expand — Method

_expand(root::RuleNode, grammar::ContextSensitiveGrammar, max_depth::Int, max_holes::Int, context::GrammarContext, iter::TopDownIterator)::Union{ExpandFailureReason, Vector{TreeConstraints}}

Recursive expand function used in multiple enumeration techniques. Expands one hole/undefined leaf of the given RuleNode tree found using the given hole heuristic. If the expansion was successful, returns a list of new trees and a list of lists of hole locations, corresponding to the holes of each newly expanded tree. Returns nothing if tree is already complete (i.e. contains no holes). Returns an empty list if the tree is partial (i.e. contains holes), but they could not be expanded because of the depth limit.

HerbSearch._find_next_complete_tree — Method

_find_next_complete_tree(grammar::ContextSensitiveGrammar, max_depth::Int, max_size::Int, pq::PriorityQueue, iter::TopDownIterator)::Union{Tuple{RuleNode, PriorityQueue}, Nothing}

Takes a priority queue and returns the smallest AST from the grammar it can obtain from the queue or by (repeatedly) expanding trees that are in the queue. Returns nothing if there are no trees left within the depth limit.

HerbSearch.best_accept — Method

best_accept(current_cost::Real, next_cost::Real, temperature::Real)

Returns true if the cost of the proposed program is smaller than the cost of the current program. Otherwise, returns false.

Arguments

current_cost::Real: the cost of the current program.

next_cost::Real: the cost of the proposed program.

temperature::Real: the temperature; not used.

HerbSearch.calculate_cost — Method

calculate_cost(iter::T, program::RuleNode) where T <: StochasticSearchIterator

Wrapper around _calculate_cost.

HerbSearch.const_temperature — Method

const_temperature(current_temperature::Real)

Returns the temperature unchanged. This function is used by Metropolis Hastings and Very Large Neighbourhood Search algorithms.

Arguments

current_temperature::Real: the current temperature of the search.

HerbSearch.constructNeighbourhood — Method

constructNeighbourhood(current_program::RuleNode, grammar::AbstractGrammar)

The neighbourhood node location is chosen at random. The dictionary is nothing.

Arguments

current_program::RuleNode: the current program.

grammar::AbstractGrammar: the grammar.

HerbSearch.constructNeighbourhoodRuleSubset — Method

constructNeighbourhoodRuleSubset(current_program::RuleNode, grammar::AbstractGrammar)

The neighbourhood node location is chosen at random. The dictionary is contains one entry with key "rule_subset" and value of type Vector{Any} being a random subset of grammar rules.

Arguments

current_program::RuleNode: the current program.

grammar::AbstractGrammar: the grammar.

HerbSearch.count_expressions — Method

count_expressions(grammar::AbstractGrammar, max_depth::Int, max_size::Int, sym::Symbol)

Counts and returns the number of possible expressions of a grammar up to max_depth with start symbol sym.

HerbSearch.count_expressions — Method

count_expressions(iter::ProgramIterator)

Counts and returns the number of possible expressions in the expression iterator. The Iterator is not modified.

HerbSearch.cross_over — Method

cross_over(::GeneticSearchIterator, parent_1::RuleNode, parent_2::RuleNode)

Combines the program from two parent individuals to create one or more offspring individuals.

HerbSearch.crossover_swap_children_1 — Method

crossover_swap_children_1(parent_1::RuleNode, parent_2::RuleNode)

Performs a random crossover of two parents of type RuleNode. The subprograms are swapped and only one altered parent program is returned.

HerbSearch.crossover_swap_children_2 — Method

crossover_swap_children_2(parent_1::RuleNode, parent_2::RuleNode)

Performs a random crossover of two parents of type RuleNode. The subprograms are swapped and both altered parent programs are returned.

HerbSearch.decreasing_temperature — Method

decreasing_temperature(percentage::Real)

Returns a function that produces a temperature decreased by percentage%. This function is used by the Simmulated Annealing algorithm.

Arguments

percentage::Real: the percentage to decrease the temperature by.

HerbSearch.default_fitness — Method

default_fitness(program, results)

Defines the default fitness function taking the program and its results. Results are a vector of tuples, where each tuple is in the form Tuple{expected_output, actual_output}. As we are looking for individuals with the highest fitness function, the error is inverted.

HerbSearch.derivation_heuristic — Method

derivation_heuristic(::TopDownIterator, nodes::Vector{RuleNode}, ::GrammarContext)::Vector{AbstractRuleNode}

Returns an ordered sublist of nodes, based on which ones are most promising to fill the hole at the given context.

nodes::Vector{RuleNode}: a list of nodes the hole can be filled with

context::GrammarContext: holds the location of the to be filled hole

HerbSearch.enumerate_neighbours_propose — Method

enumerate_neighbours_propose(enumeration_depth::Int64)

The return function is a function that produces a list with all the subprograms with depth at most enumeration_depth.

Arguments

enumeration_depth::Int64: the maximum enumeration depth.

HerbSearch.evaluate — Method

evaluate(problem::Problem{Vector{IOExample}}, expr::Any, tab::SymbolTable; allow_evaluation_errors::Bool=false)

Evaluate the expression on the examples.

Optional parameters:

- `shortcircuit` - Whether to stop evaluating after finding single example fails, to speed up the [synth](@ref) procedure. If true, the returned score is an underapproximation of the actual score.
- `allow_evaluation_errors` - Whether the search should continue if an exception is thrown in the evaluation or throw the error

Returns a score in the interval [0, 1]

HerbSearch.extractname — Method

extractname(ex)

Extracts the name of a field declaration, otherwise throws an ArgumentError. A field declaration is of the form <name>[::<type>]

HerbSearch.fitness — Method

fitness(::GeneticSearchIterator, program, results)

Assigns a numerical value (fitness score) to each individual based on how closely it meets the desired objective.

HerbSearch.get_best_program — Method

get_best_program(population::Array{RuleNode}, iter::GeneticSearchIterator)::RuleNode

Returns the best program within the population with respect to the fitness function.

HerbSearch.heuristic_leftmost — Method

heuristic_leftmost(node::AbstractRuleNode, max_depth::Int)::Union{ExpandFailureReason, HoleReference}

Defines a heuristic over holes, where the left-most hole always gets considered first. Returns a HoleReference once a hole is found. This is the default option for enumerators.

HerbSearch.heuristic_random — Method

heuristic_random(node::AbstractRuleNode, max_depth::Int)::Union{ExpandFailureReason, HoleReference}

Defines a heuristic over holes, where random holes get chosen randomly using random exploration. Returns a HoleReference once a hole is found.

HerbSearch.heuristic_rightmost — Method

heuristic_rightmost(node::AbstractRuleNode, max_depth::Int)::Union{ExpandFailureReason, HoleReference}

Defines a heuristic over holes, where the right-most hole always gets considered first. Returns a HoleReference once a hole is found.

HerbSearch.heuristic_smallest_domain — Method

heuristic_smallest_domain(node::AbstractRuleNode, max_depth::Int)::Union{ExpandFailureReason, HoleReference}

Defines a heuristic over all available holes in the unfinished AST, by considering the size of their respective domains. A domain here describes the number of possible derivations with respect to the constraints. Returns a HoleReference once a hole is found.

HerbSearch.hole_heuristic — Method

hole_heuristic(::TopDownIterator, node::AbstractRuleNode, max_depth::Int)::Union{ExpandFailureReason, HoleReference}

Defines a heuristic over holes. Returns a HoleReference once a hole is found.

HerbSearch.is_field_decl — Method

is_field_decl(ex)

Check if extractname(ex) returns a name.

HerbSearch.mean_squared_error — Method

mean_squared_error(results::AbstractVector{Tuple{<:Number,<:Number}})

Returns the mean squared error of results.

Arguments

results<:AbstractVector{<:Tuple{Number,Number}}: the vector of tuples, where each tuple is in the form Tuple{expected_output, actual_output}.

HerbSearch.misclassification — Method

misclassification(results::AbstractVector{Tuple{<:Number,<:Number}})

Returns the amount of misclassified examples, i.e. how many tuples with non-matching entries are there in results.

Arguments

results<:AbstractVector{<:Tuple{Number,Number}}: the vector of tuples, where each tuple is in the form Tuple{expected_output, actual_output}.

HerbSearch.mutate! — Function

mutate!(::GeneticSearchIterator, program::RuleNode, grammar::AbstractGrammar, max_depth::Int = 2)

Mutates the program of an invididual.

HerbSearch.mutate_random! — Function

mutate_random!(program::RuleNode, grammar::AbstractGrammar, max_depth::Int64 = 2)

Mutates the given program by inserting a randomly generated sub-program at a random location.

HerbSearch.priority_function — Method

priority_function(::BFSIterator, g::AbstractGrammar, tree::AbstractRuleNode, parent_value::Union{Real, Tuple{Vararg{Real}}})

Assigns priority such that the search tree is traversed like in a BFS manner

HerbSearch.priority_function — Method

priority_function(::DFSIterator, g::AbstractGrammar, tree::AbstractRuleNode, parent_value::Union{Real, Tuple{Vararg{Real}}})

Assigns priority such that the search tree is traversed like in a DFS manner

HerbSearch.priority_function — Method

priority_function(::MLFSIterator, g::AbstractGrammar, tree::AbstractRuleNode, parent_value::Union{Real, Tuple{Vararg{Real}}})

Calculates logit for all possible derivations for a node in a tree and returns them.

HerbSearch.priority_function — Method

priority_function(::TopDownIterator, g::AbstractGrammar, tree::AbstractRuleNode, parent_value::Union{Real, Tuple{Vararg{Real}}})

Assigns a priority value to a tree that needs to be considered later in the search. Trees with the lowest priority value are considered first.

g: The grammar used for enumeration

tree: The tree that is about to be stored in the priority queue

parent_value: The priority value of the parent PriorityQueueItem

HerbSearch.probabilistic_accept — Method

probabilistic_accept(current_cost::Real, next_cost::Real, temperature::Real)

Probabilistically decides whether to accept the new program (next) based on the ratio of costs (smaller is better) between the previous and new program. Returns True if the new program is accepted, False otherwise.

Arguments

current_cost::Real: the cost of the current program.

next_cost::Real: the cost of the proposed program.

temperature::Real: the temperature; not used.

HerbSearch.probabilistic_accept_with_temperature — Method

probabilistic_accept_with_temperature(current_cost::Real, next_cost::Real, temperature::Real)

Returns true if the cost of the proposed program is smaller than the cost of the current program. Otherwise, returns true with the probability equal to:

\[1 / (1 + exp(delta / temperature))\]

In any other case, returns false.

Arguments

current_cost::Real: the cost of the current program.

next_cost::Real: the cost of the proposed program.

temperature::Real: the temperature of the search.

HerbSearch.probabilistic_accept_with_temperature_fraction — Method

probabilistic_accept_with_temperature_fraction(current_cost::Real, program_to_consider_cost::Real, temperature::Real)

Probabilistically decides whether to accept the new program (next) based on the ratio of costs (smaller is better) between the previous and new program multiplied by the temperature. Returns True if the new program is accepted, False otherwise.

Arguments

current_cost::Real: the cost of the current program.

next_cost::Real: the cost of the proposed program.

temperature::Real: the current temperature

HerbSearch.processkwarg! — Method

processkwarg!(keywords::Vector{Expr}, ex::Union{Expr, Symbol})

Checks if ex has a default value specified, if so it returns only the field declaration, and pushes ex to keywords. Otherwise it returns ex

HerbSearch.propagate_constraints — Function

function propagate_constraints(root::AbstractRuleNode, grammar::ContextSensitiveGrammar, local_constraints::Set{LocalConstraint}, max_holes::Int, filled_hole::Union{HoleReference, Nothing}=nothing)::Tuple{PropagateResult, Set{LocalConstraint}}

Propagates a set of local constraints recursively to all children of a given root node. As propagate_constraints gets often called when a hole was just filled, filled_hole helps keeping track to propagate the constraints to relevant nodes, e.g. children of filled_hole. max_holes makes sure that max_size of Base.iterate is not violated. The function returns the PropagateResult and the set of relevant LocalConstraints.

HerbSearch.random_fill_propose — Method

random_fill_propose(current_program::RuleNode, neighbourhood_node_loc::NodeLoc, grammar::AbstractGrammar, max_depth::Int, dmap::AbstractVector{Int}, dict::Union{Nothing,Dict{String,Any}})

Returns a list with only one proposed, completely random, subprogram.

Arguments

current_program::RuleNode: the current program.

neighbourhood_node_loc::NodeLoc: the location of the program to replace.

grammar::AbstractGrammar: the grammar used to create programs.

max_depth::Int: the maximum depth of the resulting programs.

dmap::AbstractVector{Int} : the minimum possible depth to reach for each rule

dict::Dict{String, Any}: the dictionary with additional arguments; not used.

HerbSearch.select_chromosome — Method

select_chromosome(population::Array{RuleNode}, fitness_array::Array{<:Real})::RuleNode

Selects a chromosome (individual) from the population based on a fitness array. The function uses a fitness-proportionate selection strategy, often referred to as "roulette wheel" selection. Assumes fitness_array to be normalized already.

HerbSearch.select_fitness_proportional_parents — Method

select_fitness_proportional_parents(population::Array{RuleNode}, fitness_array::Array{<:Real})::Tuple{RuleNode,RuleNode}

Selects two parent chromosomes (individuals) from a population based on fitness-proportionate selection. The selected parents can be used for genetic crossover in the next steps of the algorithm.

HerbSearch.select_parents — Method

select_parents(::GeneticSearchIterator, population::Array{RuleNode}, fitness_array::Array{<:Real})

Selects two parents for the crossover.

HerbSearch.synth — Method

synth(problem::Problem, iterator::ProgramIterator; shortcircuit::Bool=true, allow_evaluation_errors::Bool=false, mod::Module=Main)::Union{Tuple{RuleNode, SynthResult}, Nothing}

Synthesize a program that satisfies the maximum number of examples in the problem. - problem - The problem definition with IO examples - iterator - The iterator that will be used - shortcircuit - Whether to stop evaluating after finding a single example that fails, to speed up the synth procedure. If true, the returned score is an underapproximation of the actual score. - allowevaluationerrors - Whether the search should crash if an exception is thrown in the evaluation - mod - A module containing definitions for the functions in the grammar that do not exist in Main

Returns a tuple of the rulenode representing the solution program and a synthresult that indicates if that program is optimal. synth uses evaluate which returns a score in the interval [0, 1] and checks whether that score reaches 1. If not it will return the best program so far, with the proper flag

HerbSearch.validate_iterator — Method

validate_iterator(iter)

Validates the parameters of the iterator

StatsBase.sample — Function

sample(root::RuleNode, typ::Symbol, grammar::AbstractGrammar, maxdepth::Int=typemax(Int))

Uniformly samples a random node from the tree limited to maxdepth.

StatsBase.sample — Function

sample(::Type{NodeLoc}, root::RuleNode, maxdepth::Int=typemax(Int))

Uniformly selects a random node in the tree no deeper than maxdepth using reservoir sampling. Returns a NodeLoc that specifies the location using its parent so that the subtree can be replaced.

StatsBase.sample — Function

sample(::Type{NodeLoc}, root::RuleNode, typ::Symbol, grammar::AbstractGrammar)

Uniformly selects a random node in the tree of a given type, specified using its parent such that the subtree can be replaced. Returns a NodeLoc.

StatsBase.sample — Function

sample(root::RuleNode, typ::Symbol, grammar::AbstractGrammar,
                      maxdepth::Int=typemax(Int))

Uniformly selects a random node of the given return type typ limited by maxdepth.

The HerbSearch package takes care of all operations related to searching for the desired program. This includes

the functionality to sample a certain program given a grammar,

the implementation of several heuristic functions,

searching for a program that satisfies the specification, and

implementations of several search algorithms in terms of how they enumerate the search space

Breadth-First Search

Depth-First Search

Metropolis Hastings

Very Large Scale Neighbourhood Search

Simulated Annealing

Genetic Search

Index

HerbConstraints.AbstractMatchNode

HerbConstraints.ComesAfter

HerbConstraints.ComesAfter

HerbConstraints.Condition

HerbConstraints.Forbidden

HerbConstraints.ForbiddenPath

HerbConstraints.GrammarContext

HerbConstraints.LocalCondition

HerbConstraints.LocalConstraint

HerbConstraints.LocalForbidden

HerbConstraints.LocalOneOf

HerbConstraints.LocalOrdered

HerbConstraints.MatchFail

HerbConstraints.MatchNode

HerbConstraints.MatchVar

HerbConstraints.OneOf

HerbConstraints.Ordered

HerbConstraints.PropagatorConstraint

HerbConstraints.RequireOnLeft

HerbCore.AbstractGrammar

HerbCore.AbstractRuleNode

HerbCore.Constraint

HerbCore.Hole

HerbCore.HoleReference

HerbCore.RuleNode

HerbCore.RuleNode

HerbCore.RuleNode

HerbGrammar.ContextSensitiveGrammar

HerbGrammar.NodeLoc

HerbGrammar.SymbolTable

HerbGrammar.SymbolTable

HerbSearch.BFSIterator

HerbSearch.DFSIterator

HerbSearch.ExpandFailureReason

HerbSearch.GeneticSearchIterator

HerbSearch.MHSearchIterator

HerbSearch.MLFSIterator

HerbSearch.PriorityQueueItem

HerbSearch.PriorityQueueItem

HerbSearch.ProgramIterator

HerbSearch.PropagateResult

HerbSearch.SASearchIterator

HerbSearch.StochasticSearchIterator

HerbSearch.SynthResult

HerbSearch.TopDownIterator

HerbSearch.VLSNSearchIterator

HerbSpecification.AbstractDependentTypeSpecification

HerbSpecification.AgdaSpecification

HerbSpecification.IOExample

HerbSpecification.MetricProblem

HerbSpecification.Problem

HerbSpecification.SMTSpecification

HerbSpecification.Trace

Base.get

Base.getindex

Base.insert!

Base.isless

Base.iterate

Base.iterate

Base.iterate

Base.iterate

Base.iterate

Base.length

Base.length

Base.rand

Base.rand

Base.rand

Base.show

Base.show

HerbConstraints._matchnode2expr

HerbConstraints._matchnode2expr

HerbConstraints._matchnode2expr

HerbConstraints._matchnode2expr

HerbConstraints._pattern_match

HerbConstraints._pattern_match

HerbConstraints._pattern_match_with_hole

HerbConstraints._pattern_match_with_hole

HerbConstraints._pattern_match_with_hole

HerbConstraints._pattern_match_with_hole

HerbConstraints._rulenode_compare

HerbConstraints._rulenode_match_with_hole

HerbConstraints.addparent!

HerbConstraints.annotation2constraint

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.copy_and_insert

HerbConstraints.make_smaller_or_equal

HerbConstraints.matchnode2expr

HerbConstraints.matchnode2expr

HerbConstraints.propagate

HerbConstraints.propagate

HerbConstraints.propagate

HerbConstraints.propagate

HerbConstraints.propagate

HerbConstraints.propagate

HerbConstraints.propagate

HerbConstraints.propagate

HerbConstraints.propagate

HerbCore.contains_hole

HerbCore.depth

HerbCore.get_node_at_location

HerbCore.get_rulesequence

HerbCore.node_depth

HerbCore.rulesoftype

HerbCore.rulesonleft

HerbCore.swap_node

HerbCore.swap_node

HerbGrammar.add_rule!

HerbGrammar.add_rule!

HerbGrammar.addconstraint!

HerbGrammar.child_types

HerbGrammar.child_types

HerbGrammar.cleanup_removed_rules!

HerbGrammar.clearconstraints!

HerbGrammar.containedin

HerbGrammar.contains_returntype

HerbGrammar.expr2csgrammar

HerbGrammar.expr2pcsgrammar

HerbGrammar.get_childtypes

HerbGrammar.get_domain

HerbGrammar.get_domain

HerbGrammar.get_node_at_location

HerbGrammar.get_rulesequence

HerbGrammar.has_children

HerbGrammar.iscomplete

HerbGrammar.iseval

HerbGrammar.iseval

HerbGrammar.iseval

HerbGrammar.isprobabilistic

HerbGrammar.isterminal

HerbGrammar.isterminal

HerbGrammar.isterminal

HerbGrammar.isvariable

HerbGrammar.isvariable

HerbGrammar.isvariable

HerbGrammar.isvariable

HerbGrammar.log_probability

HerbGrammar.max_arity

HerbGrammar.merge_grammars!

HerbGrammar.mindepth

HerbGrammar.mindepth_map

HerbGrammar.nchildren

HerbGrammar.nchildren

HerbGrammar.nonterminals

HerbGrammar.normalize!

HerbGrammar.parse_probabilistic_rule

HerbGrammar.probability

HerbGrammar.read_csg

HerbGrammar.read_pcsg

HerbGrammar.remove_rule!

HerbGrammar.return_type

HerbGrammar.return_type

HerbGrammar.root_node_loc

HerbGrammar.rulenode2expr

HerbGrammar.rulenode_log_probability

HerbGrammar.rulesoftype

HerbGrammar.rulesoftype

HerbGrammar.rulesoftype

HerbGrammar.rulesonleft

HerbGrammar.store_csg

HerbGrammar.subsequenceof

HerbGrammar.swap_node

HerbGrammar.swap_node

HerbInterpret.evaluate_program

HerbInterpret.execute_on_input

HerbInterpret.execute_on_input

HerbInterpret.execute_on_input

HerbInterpret.execute_on_input

HerbInterpret.interpret

HerbInterpret.test_all_examples

HerbInterpret.test_examples

HerbSearch._calculate_cost

HerbSearch._expand

HerbSearch._expand

HerbSearch._find_next_complete_tree

HerbSearch.best_accept

HerbSearch.calculate_cost

HerbSearch.const_temperature

HerbSearch.constructNeighbourhood

HerbSearch.constructNeighbourhoodRuleSubset

HerbSearch.count_expressions

HerbSearch.count_expressions

HerbSearch.cross_over

HerbSearch.crossover_swap_children_1

HerbSearch.crossover_swap_children_2

HerbSearch.decreasing_temperature

HerbSearch.default_fitness

HerbSearch.derivation_heuristic

HerbSearch.enumerate_neighbours_propose

HerbSearch.evaluate

HerbSearch.extractname

HerbSearch.fitness

HerbSearch.get_best_program

HerbSearch.heuristic_leftmost

HerbSearch.heuristic_random

HerbSearch.heuristic_rightmost

HerbSearch.heuristic_smallest_domain

HerbSearch.hole_heuristic

HerbSearch.is_field_decl

HerbSearch.mean_squared_error

HerbSearch.misclassification

HerbSearch.mutate!

HerbSearch.mutate_random!

HerbSearch.priority_function

HerbSearch.priority_function

HerbSearch.priority_function

HerbSearch.priority_function

HerbSearch.probabilistic_accept

HerbSearch.probabilistic_accept_with_temperature

HerbSearch.probabilistic_accept_with_temperature_fraction

HerbSearch.processkwarg!

HerbSearch.propagate_constraints

HerbSearch.random_fill_propose

HerbSearch.select_chromosome

HerbSearch.select_fitness_proportional_parents

HerbSearch.select_parents

HerbSearch.synth

HerbSearch.validate_iterator

StatsBase.sample

StatsBase.sample

StatsBase.sample

StatsBase.sample

HerbConstraints.@csgrammar_annotated

HerbGrammar.@csgrammar

HerbGrammar.@pcsgrammar

HerbSearch.@programiterator