HerbSpecification.jl Documentation

HerbSpecification.AbstractDependentTypeSpecification — Type

struct AbstractDependentTypeSpecification <: AbstractTypeSpecification

Defines a specification through dependent types. Needs a concrete type checker as oracle.

HerbSpecification.AgdaSpecification — Type

struct AgdaSpecification <: AbstractDependentTypeSpecification

Defines a specification

HerbSpecification.IOExample — Type

struct IOExample

An input-output example. in is a Dict of {Symbol,Any} where the symbol represents a variable in a program. out can be anything.

HerbSpecification.MetricProblem — Type

struct MetricProblem{T <: Vector{IOExample}}

Program synthesis problem defined by an specification and a metric. The specification has to be based on input/output examples, while the function needs to return a numerical value.

HerbSpecification.Problem — Type

struct Problem

Program synthesis problem defined by an AbstractSpecifications. Has a name and a specification of type T.

Warning

Please care that concrete Problem types with different values of T are never subtypes of each other.

HerbSpecification.SMTSpecification — Type

struct SMTSpecification <: AbstractFormalSpecification

A specification based on a logical formula defined by a SMT solver.

HerbSpecification.Trace — Type

struct Trace

A trace defining a wanted program execution for program synthesis. @TODO combine with Gen.jl

Base.getindex — Method

Base.getindex(p::Problem{Vector{IOExample}}, indices)

Overwrite Base.getindex to allow for slicing of input/output-based problems.

Index

HerbConstraints.AbstractGrammarConstraint

HerbConstraints.AbstractLocalConstraint

HerbConstraints.AbstractStateManager

HerbConstraints.Contains

HerbConstraints.ContainsSubtree

HerbConstraints.DomainRuleNode

HerbConstraints.Forbidden

HerbConstraints.ForbiddenSequence

HerbConstraints.GenericSolver

HerbConstraints.GenericSolver

HerbConstraints.GenericSolver

HerbConstraints.LessThanOrEqualHardFail

HerbConstraints.LessThanOrEqualResult

HerbConstraints.LessThanOrEqualSoftFail

HerbConstraints.LessThanOrEqualSuccess

HerbConstraints.LessThanOrEqualSuccessEquality

HerbConstraints.LessThanOrEqualSuccessLessThan

HerbConstraints.LessThanOrEqualSuccessWithHoles

HerbConstraints.LocalContains

HerbConstraints.LocalContainsSubtree

HerbConstraints.LocalContainsSubtree

HerbConstraints.LocalForbidden

HerbConstraints.LocalForbiddenSequence

HerbConstraints.LocalOrdered

HerbConstraints.LocalUnique

HerbConstraints.MakeEqualHardFail

HerbConstraints.MakeEqualResult

HerbConstraints.MakeEqualSoftFail

HerbConstraints.MakeEqualSuccess

HerbConstraints.Ordered

HerbConstraints.PatternMatchHardFail

HerbConstraints.PatternMatchResult

HerbConstraints.PatternMatchSoftFail

HerbConstraints.PatternMatchSuccess

HerbConstraints.PatternMatchSuccessWhenHoleAssignedTo

HerbConstraints.Solver

HerbConstraints.SolverState

HerbConstraints.StateHole

HerbConstraints.StateHole

HerbConstraints.StateHole

HerbConstraints.StateInt

HerbConstraints.StateIntBackup

HerbConstraints.StateManager

HerbConstraints.StateSparseSet

HerbConstraints.StateSparseSet

HerbConstraints.StateStack

HerbConstraints.StateStack

HerbConstraints.StateStack

HerbConstraints.UniformSolver

HerbConstraints.UniformSolver

HerbConstraints.Unique

HerbConstraints.VarNode

HerbCore.AbstractConstraint

HerbCore.AbstractGrammar

HerbCore.AbstractHole

HerbCore.AbstractRuleNode

HerbCore.AbstractUniformHole

HerbCore.Hole

HerbCore.HoleReference

HerbCore.RuleNode

HerbCore.RuleNode

HerbCore.RuleNode

HerbCore.UniformHole

HerbGrammar.ContextSensitiveGrammar

HerbGrammar.NodeLoc

HerbGrammar.SymbolTable

HerbSearch.AbstractBFSIterator

HerbSearch.AbstractDFSIterator

HerbSearch.AbstractMHSearchIterator

HerbSearch.AbstractSASearchIterator

HerbSearch.AbstractVLSNSearchIterator

HerbSearch.BFSIterator

HerbSearch.DFSIterator

HerbSearch.ExpandFailureReason

HerbSearch.GeneticSearchIterator

HerbSearch.MHSearchIterator

HerbSearch.MLFSIterator

HerbSearch.ProgramIterator

HerbSearch.RandomIterator

HerbSearch.SASearchIterator

HerbSearch.StochasticSearchIterator

HerbSearch.SynthResult

HerbSearch.TopDownIterator

HerbSearch.UniformIterator

HerbSearch.UniformIterator

HerbSearch.VLSNSearchIterator

HerbSpecification.AbstractDependentTypeSpecification

HerbSpecification.AgdaSpecification

HerbSpecification.IOExample

HerbSpecification.MetricProblem

HerbSpecification.Problem

HerbSpecification.SMTSpecification

HerbSpecification.Trace

Base.collect

Base.collect

Base.findall

Base.findfirst

Base.findlast

Base.get

Base.getindex

Base.getindex

Base.in

Base.in

Base.insert!

Base.isless

Base.iterate

Base.iterate

Base.iterate

Base.iterate

Base.iterate

Base.length

Base.length

Base.length

Base.length

Base.push!

Base.rand

Base.rand

Base.rand

Base.show

Base.size

Base.size

Base.sum

HerbConstraints._contains

HerbConstraints._count_occurrences

HerbConstraints._count_occurrences

HerbConstraints._count_occurrences!

HerbConstraints._exchange_positions!

HerbConstraints._update_bounds_val_removed!

HerbConstraints._update_max_val_removed!

HerbConstraints._update_min_val_removed!

HerbConstraints.annotation2constraint

HerbConstraints.are_disjoint

HerbConstraints.are_disjoint

HerbConstraints.backup!

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.contains_varnode

HerbConstraints.deactivate!

HerbConstraints.deactivate!

HerbConstraints.decrement!

HerbConstraints.fix_point!

HerbConstraints.freeze_state

HerbConstraints.get_grammar

HerbConstraints.get_grammar

HerbConstraints.get_hole_at_location

HerbConstraints.get_hole_at_location

HerbConstraints.get_intersection

HerbConstraints.get_max_depth

HerbConstraints.get_max_size

HerbConstraints.get_nodes

HerbConstraints.get_nodes_on_path

HerbConstraints.get_priority

HerbConstraints.get_starting_symbol

HerbConstraints.get_state

HerbConstraints.get_tree

HerbConstraints.get_tree

HerbConstraints.get_tree_size

HerbConstraints.get_value

HerbConstraints.increment!

HerbConstraints.is_subdomain

HerbConstraints.is_subdomain

HerbConstraints.isfeasible

HerbConstraints.isfeasible

HerbConstraints.load_state!

HerbConstraints.make_equal!

HerbConstraints.make_less_than_or_equal!

HerbConstraints.make_less_than_or_equal!

HerbConstraints.make_less_than_or_equal!

HerbConstraints.new_state!

HerbConstraints.notify_new_node

HerbConstraints.notify_new_nodes

HerbConstraints.notify_new_nodes

HerbConstraints.notify_tree_manipulation

HerbConstraints.notify_tree_manipulation

HerbConstraints.partition

HerbConstraints.pattern_match

HerbConstraints.pattern_match

HerbConstraints.pattern_match

HerbConstraints.pattern_match

HerbConstraints.pattern_match

HerbConstraints.pattern_match

HerbConstraints.post!

HerbConstraints.post!

HerbConstraints.propagate!

HerbConstraints.propagate!

HerbConstraints.propagate!

HerbConstraints.propagate!

HerbConstraints.propagate!

HerbConstraints.propagate!

HerbConstraints.propagate!

HerbConstraints.remove!

HerbConstraints.remove!

HerbConstraints.remove!

HerbConstraints.remove!

HerbConstraints.remove!

HerbConstraints.remove_above!

HerbConstraints.remove_above!

HerbConstraints.remove_above!

HerbConstraints.remove_all_but!

HerbConstraints.remove_all_but!

HerbConstraints.remove_all_but!

HerbConstraints.remove_all_but!

HerbConstraints.remove_all_but!

HerbConstraints.remove_all_but!

HerbConstraints.remove_below!

HerbConstraints.remove_below!

HerbConstraints.remove_below!

HerbConstraints.remove_node!

HerbConstraints.restore!

HerbConstraints.restore!

HerbConstraints.restore!

HerbConstraints.save_state!

HerbConstraints.save_state!

HerbConstraints.save_state!

HerbConstraints.schedule!

HerbConstraints.set_infeasible!

HerbConstraints.set_infeasible!

HerbConstraints.set_value!

HerbConstraints.shouldschedule

HerbConstraints.shouldschedule

HerbConstraints.simplify_hole!

HerbConstraints.substitute!

HerbCore.contains_hole

HerbCore.contains_hole

HerbCore.contains_index

HerbCore.contains_nonuniform_hole

HerbCore.depth

HerbCore.get_children

HerbCore.get_node_at_location

HerbCore.get_node_at_location

HerbCore.get_node_at_location

HerbCore.get_node_at_location

HerbCore.get_path

HerbCore.get_path

HerbCore.get_path

HerbCore.get_rule

HerbCore.get_rule

HerbCore.get_rulesequence

HerbCore.hasdynamicvalue

HerbCore.have_same_shape

HerbCore.is_domain_valid

HerbCore.isfilled

HerbCore.isfilled

HerbCore.issame

HerbCore.isuniform

HerbCore.node_depth

HerbCore.number_of_holes

HerbCore.rulesoftype

HerbCore.rulesoftype

HerbCore.rulesonleft

HerbCore.swap_node

HerbCore.swap_node

HerbCore.update_rule_indices!

HerbCore.update_rule_indices!

HerbCore.update_rule_indices!

HerbCore.update_rule_indices!

HerbCore.update_rule_indices!

HerbCore.update_rule_indices!

HerbCore.update_rule_indices!

HerbCore.update_rule_indices!

HerbCore.update_rule_indices!

HerbCore.update_rule_indices!

HerbCore.update_rule_indices!

HerbCore.update_rule_indices!

HerbCore.update_rule_indices!

HerbCore.update_rule_indices!

HerbCore.update_rule_indices!

HerbCore.update_rule_indices!

HerbCore.update_rule_indices!

HerbCore.update_rule_indices!

HerbCore.update_rule_indices!

HerbCore.update_rule_indices!

HerbCore.update_rule_indices!

HerbCore.update_rule_indices!

HerbCore.update_rule_indices!

HerbCore.update_rule_indices!

HerbCore.update_rule_indices!

HerbCore.update_rule_indices!

HerbCore.update_rule_indices!

HerbCore.update_rule_indices!

HerbCore.update_rule_indices!

HerbCore.update_rule_indices!

HerbCore.update_rule_indices!

HerbCore.update_rule_indices!

HerbCore.update_rule_indices!

HerbGrammar.add_rule!

HerbGrammar.add_rule!

HerbGrammar.add_rule!

HerbGrammar.addconstraint!

HerbGrammar.child_types

HerbGrammar.child_types

HerbGrammar.cleanup_removed_rules!

HerbGrammar.clearconstraints!

HerbGrammar.containedin

HerbGrammar.contains_returntype

HerbGrammar.expr2csgrammar

HerbGrammar.expr2pcsgrammar

HerbGrammar.expr2rulenode

HerbGrammar.expr2rulenode

HerbGrammar.expr2rulenode

HerbGrammar.expr2rulenode

HerbGrammar.get_childtypes

HerbGrammar.get_domain

HerbGrammar.get_domain

HerbGrammar.get_rulesequence

HerbGrammar.grammar2symboltable

HerbGrammar.init_probabilities!

HerbGrammar.iscomplete

HerbGrammar.iseval

HerbGrammar.iseval

HerbGrammar.iseval

HerbGrammar.isprobabilistic

HerbGrammar.isterminal

HerbGrammar.isterminal

HerbGrammar.isterminal

HerbGrammar.isvariable

HerbGrammar.isvariable

HerbGrammar.isvariable

HerbGrammar.isvariable

HerbGrammar.log_probability

HerbGrammar.max_arity

HerbGrammar.max_rulenode_log_probability

HerbGrammar.merge_grammars!

HerbGrammar.mindepth

HerbGrammar.mindepth_map

HerbGrammar.nchildren

HerbGrammar.nchildren

HerbGrammar.nonterminals

HerbGrammar.normalize!

HerbGrammar.parse_probabilistic_rule

HerbGrammar.probability

HerbGrammar.read_csg

HerbGrammar.read_pcsg

HerbGrammar.remove_rule!

HerbGrammar.return_type

HerbGrammar.return_type

HerbGrammar.return_type

HerbGrammar.root_node_loc

HerbGrammar.rulenode2expr

HerbGrammar.rulenode_log_probability

HerbGrammar.rulesonleft

HerbGrammar.store_csg

HerbGrammar.subsequenceof

HerbGrammar.swap_node

HerbGrammar.swap_node

HerbInterpret.evaluate_program

HerbInterpret.execute_on_input

HerbInterpret.execute_on_input

HerbInterpret.execute_on_input

HerbInterpret.execute_on_input

HerbInterpret.interpret

HerbSearch._calculate_cost

HerbSearch._find_next_complete_tree

HerbSearch.accept

HerbSearch.best_accept

HerbSearch.calculate_cost

HerbSearch.const_temperature

HerbSearch.constructNeighbourhood

HerbSearch.constructNeighbourhoodRuleSubset

HerbSearch.cross_over

HerbSearch.crossover_swap_children_1

HerbSearch.crossover_swap_children_2

HerbSearch.decreasing_temperature

HerbSearch.default_fitness

HerbSearch.derivation_heuristic

HerbSearch.derivation_heuristic

HerbSearch.derivation_heuristic

HerbSearch.enumerate_neighbours_propose

HerbSearch.evaluate

HerbSearch.extract_name_from_argument

HerbSearch.fitness

HerbSearch.generate_branches

HerbSearch.get_best_program

HerbSearch.heuristic_leftmost

HerbSearch.heuristic_random

HerbSearch.heuristic_rightmost

HerbSearch.heuristic_smallest_domain

HerbSearch.hole_heuristic

HerbSearch.is_field_decl

HerbSearch.is_kwdef

HerbSearch.mean_squared_error

HerbSearch.misclassification

HerbSearch.mutate!

HerbSearch.mutate_random!

HerbSearch.neighbourhood

HerbSearch.next_solution!

HerbSearch.priority_function

HerbSearch.priority_function

HerbSearch.priority_function

HerbSearch.priority_function

HerbSearch.probabilistic_accept

HerbSearch.probabilistic_accept_with_temperature

HerbSearch.probabilistic_accept_with_temperature_fraction

HerbSearch.processkwarg!

HerbSearch.propose

HerbSearch.random_fill_propose

HerbSearch.select_chromosome

HerbSearch.select_fitness_proportional_parents

HerbSearch.select_parents

HerbSearch.set_stateholes!

HerbSearch.synth

HerbSearch.temperature

HerbSearch.validate_iterator

StatsBase.sample

StatsBase.sample

StatsBase.sample

StatsBase.sample

HerbConstraints.@csgrammar_annotated

HerbCore.@rulenode

HerbGrammar.@cfgrammar

HerbGrammar.@csgrammar

HerbGrammar.@pcsgrammar

HerbSearch.@programiterator