HerbConstraints.jl Documentation

HerbConstraints.AbstractGrammarConstraint — Type

abstract type AbstractGrammarConstraint <: AbstractConstraint

Abstract type representing all user-defined constraints. Each grammar constraint has a related AbstractLocalConstraint that is responsible for propagating the constraint at a specific location in the tree. Grammar constraints should implement on_new_node to post a AbstractLocalConstraint at that new node

HerbConstraints.AbstractLocalConstraint — Type

abstract type AbstractLocalConstraint <: AbstractConstraint

Abstract type representing all local constraints. Each local constraint contains a path that points to a specific location in the tree at which the constraint applies.

Each local constraint should implement a propagate!-function. Inside the propagate! function, the constraint can use the following solver functions:

remove!: Elementary tree manipulation. Removes a value from a domain. (other tree manipulations are: remove_above!, remove_below!, remove_all_but!)

deactivate!: Prevent repropagation. Call this as soon as the constraint is satisfied.

set_infeasible!: Report a non-trivial inconsistency. Call this if the constraint can never be satisfied. An empty domain is considered a trivial inconsistency, such inconsistencies are already handled by tree manipulations.

isfeasible: Check if the current tree is still feasible. Return from the propagate function, as soon as infeasibility is detected.

HerbConstraints.AbstractStateManager — Type

Manages all changes made to StateInts using StateIntBackups. Support the following functions:

StateInt Creates a new stateful integer

save_state! Creates a checkpoint for all stateful integers

restore! Restores the values to the latest checkpoint

HerbConstraints.Contains — Type

Contains <: AbstractGrammarConstraint This [AbstractGrammarConstraint] enforces that a given rule appears in the program tree at least once.

HerbConstraints.ContainsSubtree — Type

ContainsSubtree <: AbstractGrammarConstraint

This [AbstractGrammarConstraint] enforces that a given subtree appears in the program tree at least once.

!!! warning: This constraint can only be propagated by the UniformSolver

HerbConstraints.DomainRuleNode — Type

struct DomainRuleNode <: AbstractRuleNode

Matches any 1 rule in its domain. Example usage:

DomainRuleNode(Bitvector((0, 0, 1, 1)), [RuleNode(1), RuleNode(1)])

This matches RuleNode(3, [RuleNode(1), RuleNode(1)]) and RuleNode(4, [RuleNode(1), RuleNode(1)]) and UniformHole({3, 4}, [RuleNode(1), RuleNode(1)])

HerbConstraints.Forbidden — Type

Forbidden <: AbstractGrammarConstraint

This [AbstractGrammarConstraint] forbids any subtree that matches the pattern given by tree to be generated. A pattern is a tree of AbstractRuleNodes. Such a node can either be a RuleNode, which contains a rule index corresponding to the rule index in the AbstractGrammar and the appropriate number of children, similar to RuleNodes. It can also contain a VarNode, which contains a single identifier symbol. A VarNode can match any subtree, but if there are multiple instances of the same variable in the pattern, the matched subtrees must be identical. Any rule in the domain that makes the match attempt successful is removed.

For example, consider the tree 1(a, 2(b, 3(c, 4)))):

Forbidden(RuleNode(3, [RuleNode(5), RuleNode(4)])) forbids c to be filled with 5.

Forbidden(RuleNode(3, [VarNode(:v), RuleNode(4)])) forbids c to be filled, since a [VarNode] can match any rule, thus making the match attempt successful for the entire domain of c. Therefore, this tree invalid.

Forbidden(RuleNode(3, [VarNode(:v), VarNode(:v)])) forbids c to be filled with 4, since that would make both assignments to v equal, which causes a successful match.

HerbConstraints.ForbiddenSequence — Type

ForbiddenPath <: AbstractGrammarConstraint

This [AbstractGrammarConstraint] forbids the given sequence of rule nodes. Sequences are strictly vertical and may include gaps. Consider the tree 1(a, 2(b, 3(c, d)))):

[2, 3, d] is a sequence

[1, 3, d] is a sequence

[3, c, d] is not a sequence

Examples:

ForbiddenSequence([3, 4]) enforces that rule 4 cannot be applied at c or d.

ForbiddenSequence([1, 2, 4]) enforces that rule 4 cannot be applied at b, c or d.

ForbiddenSequence([1, 4]) enforces that rule 4 cannot be applied anywhere.

If any of the rules in ignore_if appears in the sequence, the constraint is ignored. Suppose the forbidden sequence = [1, 2, 3] and ignore_if = [99] Consider the following paths from the root:

[1, 2, 2, 3] is forbidden, as the sequence does not contain 99

[1, 99, 2, 3] is NOT forbidden, as the sequence does contain 99

[1, 99, 1, 2, 3] is forbidden, as there is a subsequence that does not contain 99

HerbConstraints.GenericSolver — Type

GenericSolver

Maintains a feasible partial program in a SolverState. A ProgramIterator may manipulate the partial tree with the following tree manipulations:

substitute!

remove!

remove_below!

remove_above!

remove_all_but!

Each SolverState holds an independent propagation program. Program iterators can freely move back and forth between states using:

new_state!

save_state!

load_state!

HerbConstraints.GenericSolver — Method

GenericSolver(grammar::AbstractGrammar, init_node::AbstractRuleNode)

Constructs a new solver, with an initial state of the provided AbstractRuleNode.

HerbConstraints.GenericSolver — Method

GenericSolver(grammar::AbstractGrammar, sym::Symbol)

Constructs a new solver, with an initial state using starting symbol sym

HerbConstraints.LessThanOrEqualHardFail — Type

struct LessThanOrEqualHardFail <: LessThanOrEqualResult end

node1 > node2 is guaranteed under all possible assignments of the holes involved.

HerbConstraints.LessThanOrEqualResult — Type

abstract type LessThanOrEqualResult end

A result of the less_than_or_equal function. Can be one of 3 cases:

LessThanOrEqualSuccess

LessThanOrEqualHardFail

LessThanOrEqualSoftFail

HerbConstraints.LessThanOrEqualSoftFail — Type

struct LessThanOrEqualSoftFail <: LessThanOrEqualResult

node1 <= node2 and node1 > node2 are both possible depending on the assignment of hole1 and hole2. Includes two cases:

hole2::AbstractHole: A failed AbstractHole-AbstractHole comparison. (e.g. AbstractHole(BitVector((1, 0, 1))) vs AbstractHole(BitVector((0, 1, 1))))

hole2::Nothing: A failed AbstractHole-RuleNode comparison. (e.g. AbstractHole(BitVector((1, 0, 1))) vs RuleNode(2))

HerbConstraints.LessThanOrEqualSuccess — Type

abstract type LessThanOrEqualSuccess <: LessThanOrEqualResult

node1 <= node2 is guaranteed under all possible assignments of the holes involved. The strictness of a LessThanOrEqualSuccess is specified by 1 of 2 concrete cases:

LessThanOrEqualSuccessLessThan: node1 < node2

LessThanOrEqualSuccessEquality: node1 == node2

LessThanOrEqualSuccessWithHoles: node1 <= node2. Unable to specific.

HerbConstraints.LessThanOrEqualSuccessEquality — Type

struct LessThanOrEqualSuccessEquality <: LessThanOrEqualSuccess end

node1 == node2 is guaranteed under all possible assignments of the holes involved.

HerbConstraints.LessThanOrEqualSuccessLessThan — Type

struct LessThanOrEqualSuccessEquality <: LessThanOrEqualSuccess end

node1 < node2 is guaranteed under all possible assignments of the holes involved.

HerbConstraints.LessThanOrEqualSuccessWithHoles — Type

struct LessThanOrEqualSuccessWithHoles <: LessThanOrEqualSuccess end

node1 <= node2 is guaranteed under all possible assignments of the holes involved. Because of the holes involved, it is not possible to specify '<' or '=='.

HerbConstraints.LocalContains — Type

LocalContains

Enforces that a given rule appears at or below the given path at least once.

HerbConstraints.LocalContainsSubtree — Type

LocalContains

Enforces that a given tree appears at or below the given path at least once.

!!! warning: This is a stateful constraint can only be propagated by the UniformSolver. The indices and candidates fields should not be set by the user.

HerbConstraints.LocalContainsSubtree — Method

LocalContainsSubtree(path::Vector{Int}, tree::AbstractRuleNode)

Enforces that a given tree appears at or below the given path at least once.

HerbConstraints.LocalForbidden — Type

LocalForbidden

Forbids the a subtree that matches the tree to be generated at the location provided by the path. Use a Forbidden constraint for enforcing this throughout the entire search space.

HerbConstraints.LocalForbiddenSequence — Type

LocalForbiddenSequence <: AbstractLocalConstraint

Forbids the given sequence of rule nodes ending at the node at the path. If any of the rules in ignore_if appears in the sequence, the constraint is ignored.

HerbConstraints.LocalOrdered — Type

Enforces an order over two or more subtrees that fill the variables specified in order when the pattern is applied at the location given by path. Use an Ordered constraint for enforcing this throughout the entire search space.

HerbConstraints.LocalUnique — Type

LocalUnique <: AbstractLocalConstraint

Enforces that a given rule appears at or below the given path at most once. In case of the UniformSolver, cache the list of holes, since no new holes can appear.

HerbConstraints.MakeEqualHardFail — Type

struct MakeEqualHardFail <: MakeEqualResult end

node1 != node2 is guaranteed under all possible assignments of the holes involved.

HerbConstraints.MakeEqualResult — Type

abstract type MakeEqualResult end

A result of the make_equal! function. Can be one of 3 cases:

MakeEqualSuccess

MakeEqualHardFail

MakeEqualSoftFail

HerbConstraints.MakeEqualSoftFail — Type

struct MakeEqualSoftFail <: MakeEqualResult end

Making node1 == node2 is ambiguous. Examples:

RuleNode(1, [Hole({1, 2, 3})]) == RuleNode(1, [VarNode(:a)]). The hole can be filled with any rule.

Hole({1, 2, 3}) == DomainRuleNode({1, 2, 3}). The hole can be filled with any rule.

HerbConstraints.MakeEqualSuccess — Type

struct MakeEqualSuccess <: MakeEqualResult end

node1 == node2 is guaranteed under all possible assignments of the holes involved.

HerbConstraints.Ordered — Type

Ordered <: AbstractGrammarConstraint

A AbstractGrammarConstraint that enforces a specific order in MatchVar assignments in the pattern defined by tree. Nodes in the pattern can either be a RuleNode, which contains a rule index corresponding to the rule index in the AbstractGrammar and the appropriate number of children. It can also contain a VarNode, which contains a single identifier symbol. A VarNode can match any subtree. If there are multiple instances of the same variable in the pattern, the matched subtrees must be identical.

The order defines an order between the variable assignments. For example, if the order is [x, y], the constraint will require the assignment to x to be less than or equal to the assignment to y. The order is recursively defined by RuleNode indices. For more information, see Base.isless(rn₁::AbstractRuleNode, rn₂::AbstractRuleNode).

For example, consider the tree 1(a, 2(b, 3(c, 4)))):

Ordered(RuleNode(3, [VarNode(:v), VarNode(:w)]), [:v, :w]) removes every rule with an index of 5 or greater from the domain of c, since that would make the index of the assignment to v greater than the index of the assignment to w, violating the order.

Ordered(RuleNode(3, [VarNode(:v), VarNode(:w)]), [:w, :v]) removes every rule with an index of 4 or less from the domain of c, since that would make the index of the assignment to v less than the index of the assignment to w, violating the order.

HerbConstraints.PatternMatchHardFail — Type

The pattern is not matched and can never be matched by filling in holes

HerbConstraints.PatternMatchResult — Type

abstract type PatternMatchResult end

A result of the pattern_match function. Can be one of 4 cases:

PatternMatchSuccess

PatternMatchSuccessWhenHoleAssignedTo

PatternMatchHardFail

PatternMatchSoftFail

HerbConstraints.PatternMatchSoftFail — Type

The pattern can still be matched in a non-trivial way. Includes two cases:

multiple holes are involved. this result stores a reference to one of them

a single hole is involved, but needs to be filled with a node of size >= 2

HerbConstraints.PatternMatchSuccess — Type

The pattern is exactly matched and does not involve any holes at all

HerbConstraints.PatternMatchSuccessWhenHoleAssignedTo — Type

The pattern can be matched when the hole is filled with any of the given ind(s).

HerbConstraints.Solver — Type

abstract type Solver

Abstract constraint solver. Each solver should have at least the following fields:

statistics::SolverStatistics

fix_point_running::Bool

schedule::PriorityQueue{AbstractLocalConstraint, Int}

Each solver should implement at least:

post!

get_tree

get_grammar

set_infeasible!

isfeasible

HerbCore.get_node_at_location

get_hole_at_location

notify_tree_manipulation

deactivate!

HerbConstraints.SolverState — Type

mutable struct SolverState

A state to be solved by the GenericSolver. A state contains of:

tree: A partial AST

active_constraints: The local constraints that apply to this tree. These constraints are enforced each time the tree is modified.

isfeasible: Flag to indicate if this state is still feasible. When a propagator spots an inconsistency, this field will be set to false. Tree manipulations and further propagations are not allowed on infeasible states

HerbConstraints.SolverStatistics — Type

Temporary struct to track! the number of several function calls centered around the Solver

HerbConstraints.StateHole — Type

StateHole <: AbstractUniformHole

StateHoles are uniform holes used by the UniformSolver. Domain manipulations are tracked for backpropagation.

domain: A StateSparseSet representing the rule nodes this hole can take. If size(domain) == 1, this hole should act like a RuleNode

children: The children of this hole in the expression tree.

HerbConstraints.StateHole — Method

Converts a RuleNode to a StateHole

HerbConstraints.StateHole — Method

Converts a UniformHole to a StateHole

HerbConstraints.StateInt — Type

Stateful integer that can be saved and restored by the StateManager. Supports the following functions:

get_value

set_value!

increment!

decrement!

HerbConstraints.StateIntBackup — Type

Backup entry for the given StateInt

HerbConstraints.StateManager — Type

Manages all changes made to StateInts using StateIntBackups

HerbConstraints.StateSparseSet — Method

Converts a BitVector domain representation to a StateSparseSet Example:

set = StateSparseSet(sm, BitVector((1, 1, 0, 0, 1, 0, 0))) #{1, 2, 5}

HerbConstraints.StateSparseSet — Method

Create a new StateSparseSet with values [1, 2, ..., n]

HerbConstraints.StateStack — Type

Simple stack that can only increase in size. Supports backtracking by decreasing the size to the saved size.

HerbConstraints.StateStack — Method

function StateStack{T}(sm::AbstractStateManager) where T

Create an empty StateStack supporting elements of type T

HerbConstraints.StateStack — Method

function StateStack{T}(sm::AbstractStateManager, vec::Vector{T}) where T

Create a StateStack for the provided vec

HerbConstraints.UniformSolver — Type

A DFS-based solver that uses StateHoles that support backtracking.

HerbConstraints.UniformSolver — Method

UniformSolver(grammar::AbstractGrammar, fixed_shaped_tree::AbstractRuleNode)

HerbConstraints.Unique — Type

Unique <: AbstractGrammarConstraint

This [AbstractGrammarConstraint] enforces that a given rule appears in the program tree at most once.

HerbConstraints.VarNode — Type

struct VarNode <: AbstractRuleNode

Matches any subtree and assigns it to a variable name. The LocalForbidden constraint will not match if identical variable symbols match to different trees. Example usage:

RuleNode(3, [VarNode(:x), VarNode(:x)])

This matches RuleNode(3, [RuleNode(1), RuleNode(1)]), RuleNode(3, [RuleNode(2), RuleNode(2)]), etc. but also larger subtrees such as RuleNode(3, [RuleNode(4, [RuleNode(1)]), RuleNode(4, [RuleNode(1)])])

HerbConstraints.@csgrammar_annotated — Macro

@csgrammar_annotated Define an annotated grammar and return it as a ContextSensitiveGrammar. Allows for adding optional annotations per rule. As well as that, allows for adding optional labels per rule, which can be referenced in annotations. Syntax is backwards-compatible with @csgrammar. Examples:

g₁ = @csgrammar_annotated begin
    Element = 1
    Element = x
    Element = Element + Element := commutative
    Element = Element * Element := (commutative, transitive)
end

g₁ = @csgrammar_annotated begin
    Element = 1
    Element = x
    Element = Element + Element := forbidden_path([3, 1])
    Element = Element * Element := (commutative, transitive)
end

g₁ = @csgrammar_annotated begin
    one::            Element = 1
    variable::       Element = x
    addition::       Element = Element + Element := (
                                                       commutative,
                                                       transitive,
                                                       forbidden_path([:addition, :one]) || forbidden_path([:one, :variable])
                                                    )
    multiplication:: Element = Element * Element := (commutative, transitive)
end

Base.collect — Method

function Base.collect(stack::StateStack)

Return the internal Vector representation of the stack. !!! warning: The returned vector is read-only.

Base.findall — Method

Returns all elements in the set.

Base.findfirst — Method

Returns the minimum value in the set. This function name is used instead of min to allow code reuse for domains of type BitVector and StateSparseSet.

Base.findlast — Method

Returns the maximum value in the set. This function name is used instead of min to allow code reuse for domains of type BitVector and StateSparseSet.

Base.getindex — Method

Checks if value val is in StateSparseSet s. !!! warning: This allows a StateSparseSet to be used as if it were a BitVector representation of a set

Base.in — Method

Checks if value val is in StateSparseSet s.

Base.in — Method

function Base.in(stack::StateStack, value)::Bool

Checks whether the value is in the stack.

Base.length — Method

Returns the number of values in the StateSparseSet.

Base.push! — Method

function Base.push!(stack::StateStack, item)

Place an item on top of the stack.

Base.show — Method

Pretty print the StateSparseSet.

Base.size — Method

Returns the number of values in the StateSparseSet.

Base.size — Method

function Base.size(stack::StateStack)

Get the current size of the stack.

Base.sum — Method

Returns the number of values in the StateSparseSet. !!! warning: This is not actually the sum of the set. It is the length of the set. This allows a StateSparseSet to be used as if it were a BitVector representation of a set

HerbConstraints._contains — Method

_contains(node::AbstractRuleNode, rule::Int)::Bool

Recursive helper function for the LocalContains constraint Returns one of the following:

true, if the node does contains the rule

false, if the node does not contain the rule

Vector{AbstractHole}, if the node contains the rule if one the holes gets filled with the target rule

HerbConstraints._count_occurrences! — Method

function _count_occurrences!(node::AbstractRuleNode, rule::Int, holes::Vector{AbstractHole})::Int

Recursive helper function for the LocalUnique constraint. Returns the number of certain occurrences of the rule in the tree. All holes that potentially can hold the target rule are stored in the holes vector.

!!! warning: Stops counting if the rule occurs more than once. Counting beyond 2 is not needed for LocalUnique.

HerbConstraints._count_occurrences — Method

function _count_occurrences(rule::Int, node::AbstractRuleNode)::Int

Recursively counts the number of occurrences of the rule in the node.

HerbConstraints._count_occurrences — Method

function _count_occurrences(holes::Vector{AbstractHole}, rule::Int)

Counts the occurences of the rule in the cached list of holes.

!!! warning: Stops counting if the rule occurs more than once. Counting beyond 2 is not needed for LocalUnique.

HerbConstraints._exchange_positions! — Method

Exchanges the positions in the internal representation of the StateSparseSet.

HerbConstraints._update_bounds_val_removed! — Method

This function should be called whenever the minimum or maximum value from the set might have been removed. The minimum and maximum value of the set will be updated to the actual bounds of the set.

HerbConstraints._update_max_val_removed! — Method

This function should be called whenever the maximum value from the set might have been removed. The maximum value of the set will be updated to the actual maximum of the set.

HerbConstraints._update_min_val_removed! — Method

This function should be called whenever the minimum value from the set might have been removed. The minimum value of the set will be updated to the actual minimum of the set.

HerbConstraints.annotation2constraint — Method

Converts an annotation to a constraint. commutative: creates an Ordered constraint transitive: creates an (incorrect) Forbidden constraint forbidden_path(path::Vector{Union{Symbol, Int}}): creates a ForbiddenPath constraint with the original rule included ... || ...: creates a OneOf constraint (also works with ... || ... || ... et cetera, though not very performant)

HerbConstraints.are_disjoint — Method

are_disjoint(domain1::BitVector, domain2::BitVector)::Bool

Returns true if there is no overlap in values between domain1 and domain2

HerbConstraints.are_disjoint — Method

are_disjoint(set1::StateSparseSet, set2::StateSparseSet)

Returns true if there is no overlap in values between set1 and set2

HerbConstraints.backup! — Method

Should be called whenever the state of a StateInt is modified. Creates a StateIntBackup for the given StateInt. Only backup the value if this integer has not been stored during this state before Example usecase:

a = StateInt(sm, 10)
save_state!(sm)
set_value!(a, 9) #backup value 10
set_value!(a, 8) #no need to backup again
set_value!(a, 3) #no need to backup again
restore!(sm) #restores a to value 10

HerbConstraints.check_tree — Method

check_tree(c::Contains, tree::AbstractRuleNode)::Bool

Checks if the given AbstractRuleNode tree abides the Contains constraint.

HerbConstraints.check_tree — Method

check_tree(c::ContainsSubtree, tree::AbstractRuleNode)::Bool

Checks if the given AbstractRuleNode tree abides the ContainsSubtree constraint.

HerbConstraints.check_tree — Method

check_tree(c::Forbidden, tree::AbstractRuleNode)::Bool

Checks if the given AbstractRuleNode tree abides the Forbidden constraint.

HerbConstraints.check_tree — Method

check_tree(c::ForbiddenSequence, tree::AbstractRuleNode; sequence_started=false)::Bool

Checks if the given AbstractRuleNode tree abides the ForbiddenSequence constraint.

HerbConstraints.check_tree — Method

check_tree(c::Ordered, tree::AbstractRuleNode)::Bool

Checks if the given AbstractRuleNode tree abides the Ordered constraint.

HerbConstraints.check_tree — Method

function check_tree(c::Unique, tree::AbstractRuleNode)::Bool

Checks if the given AbstractRuleNode tree abides the Unique constraint.

HerbConstraints.contains_varnode — Method

contains_varnode(rn::AbstractRuleNode, name::Symbol)

Checks if an AbstractRuleNode tree contains a VarNode with the given name.

HerbConstraints.deactivate! — Method

deactivate!(solver::GenericSolver, constraint::AbstractLocalConstraint)

Function that should be called whenever the constraint is already satisfied and never has to be repropagated.

HerbConstraints.deactivate! — Method

deactivate!(solver::UniformSolver, constraint::AbstractLocalConstraint)

Function that should be called whenever the constraint is already satisfied and never has to be repropagated.

HerbConstraints.decrement! — Method

Decrease the value of the integer by 1

HerbConstraints.fix_point! — Method

fix_point!(solver::Solver)

Propagate constraints in the current state until no further dedecutions can be made

HerbConstraints.freeze_state — Method

freeze_state(hole::StateHole)::RuleNode

Converts a [StateHole])(@ref) to a [RuleNode]@(ref). The hole and its children are assumed to be filled.

HerbConstraints.get_grammar — Method

function get_grammar(solver::GenericSolver)::AbstractGrammar

Get the grammar.

HerbConstraints.get_grammar — Method

function get_grammar(solver::UniformSolver)::AbstractGrammar

Get the grammar.

HerbConstraints.get_hole_at_location — Method

get_hole_at_location(solver::GenericSolver, location::Vector{Int})::AbstractHole

Get the node at path location and assert it is a AbstractHole.

HerbConstraints.get_hole_at_location — Method

get_hole_at_location(solver::UniformSolver, path::Vector{Int})

Get the hole that is located at the provided path.

HerbConstraints.get_intersection — Method

get_intersection(domain1::BitVector, domain2::BitVector)::Bool

Returns all the values that are in both domain1 and domain2

HerbConstraints.get_max_depth — Method

function get_max_depth(solver::GenericSolver)::SolverState

Get the maximum depth of the tree.

HerbConstraints.get_max_size — Method

function get_max_depth(solver::GenericSolver)::SolverState

Get the maximum number of AbstractRuleNodes allowed inside the tree.

HerbConstraints.get_nodes — Method

get_nodes(solver)

Return an iterator over all nodes in the tree

HerbConstraints.get_nodes_on_path — Method

function get_nodes_on_path(root::AbstractRuleNode, path::Vector{Int})::Vector{AbstractRuleNode}

Gets a list of nodes on the path, starting (and including) the root.

HerbConstraints.get_priority — Method

function get_priority(::AbstractLocalConstraint)

Used to determine which constraint to propagate first in fix_point!. Constraints with fast propagators and/or strong inference should be propagated first.

HerbConstraints.get_starting_symbol — Method

function get_starting_symbol(solver::GenericSolver)::Symbol

Get the symbol from the solver.

HerbConstraints.get_state — Method

function get_state(solver::GenericSolver)::SolverState

Get the current [SolverState]@(ref) of the solver.

HerbConstraints.get_tree — Method

function get_tree(solver::GenericSolver)::AbstractRuleNode

Returns the number of AbstractRuleNodes in the tree.

HerbConstraints.get_tree — Method

function get_tree(solver::UniformSolver)::AbstractRuleNode

Get the root of the tree. This remains the same instance throughout the entire search.

HerbConstraints.get_tree_size — Method

function get_tree_size(solver::GenericSolver)::Int

Returns the number of AbstractRuleNodes in the tree.

HerbConstraints.get_value — Method

Get the value of the stateful integer

HerbConstraints.increment! — Method

Increase the value of the integer by 1

HerbConstraints.is_subdomain — Method

is_subdomain(specific_tree::AbstractRuleNode, general_tree::AbstractRuleNode)

Checks if the specific_tree can be obtained by repeatedly removing values from the general_tree

HerbConstraints.is_subdomain — Method

 is_subdomain(subdomain::BitVector, domain::BitVector)
 is_subdomain(subdomain::StateSparseSet, domain::BitVector)

Checks if subdomain is a subdomain of domain. Example: [0, 0, 1, 0] is a subdomain of [0, 1, 1, 1]

HerbConstraints.isfeasible — Method

isfeasible(solver::GenericSolver)

Returns true if no inconsistency has been detected. Used in several ways:

Iterators should check for infeasibility to discard infeasible states

After any tree manipulation with the possibility of an inconsistency (e.g. remove_below!, remove_above!, remove!)

fix_point! should check for infeasibility to clear its schedule and return

Some GenericSolver functions assert a feasible state for debugging purposes @assert isfeasible(solver)

Some GenericSolver functions have a guard that skip the function on an infeasible state: if !isfeasible(solver) return end

HerbConstraints.isfeasible — Method

isfeasible(solver::UniformSolver)

Returns true if no inconsistency has been detected.

HerbConstraints.load_state! — Method

load_state!(solver::GenericSolver, state::SolverState)

Overwrites the current state with the given state

HerbConstraints.make_equal! — Method

function make_equal!(solver::Solver, node1::AbstractRuleNode, node2::AbstractRuleNode)::MakeEqualResult

Tree manipulation that enforces node1 == node2 if unambiguous.

HerbConstraints.make_less_than_or_equal! — Method

function make_less_than_or_equal!(h1::Union{RuleNode, AbstractHole}, h2::Union{RuleNode, AbstractHole}, guards::Vector{Tuple{AbstractHole, Int}})::LessThanOrEqualResult

Helper function that keeps track of the guards

HerbConstraints.make_less_than_or_equal! — Method

function make_less_than_or_equal!(h1::Union{RuleNode, AbstractHole}, h2::Union{RuleNode, AbstractHole})::LessThanOrEqualResult

Ensures that n1<=n2 by removing impossible values from holes. Returns one of the following results:

LessThanOrEqualSuccess. When [n1<=n2].

LessThanOrEqualHardFail. When [n1>n2] or when the solver state is infeasible.

LessThanOrEqualSoftFail. When no further deductions can be made, but [n1<=n2] and [n1>n2] are still possible.

HerbConstraints.make_less_than_or_equal! — Method

function make_less_than_or_equal!(solver::Solver, nodes1::Vector{AbstractRuleNode}, nodes2::Vector{AbstractRuleNode}, guards::Vector{Tuple{AbstractHole, Int}})::LessThanOrEqualResult

Helper function that tiebreaks on children.

HerbConstraints.new_state! — Method

new_state!(solver::GenericSolver, tree::AbstractRuleNode)

Overwrites the current state and propagates constraints on the tree from the ground up

HerbConstraints.notify_new_node — Method

notify_new_node(solver::GenericSolver, event_path::Vector{Int})

Notify all constraints that a new node has appeared at the event_path by calling their respective on_new_node function.

Warning

This does not notify the solver about nodes below the event_path. In that case, call notify_new_nodes instead.

HerbConstraints.notify_new_nodes — Method

notify_new_nodes(solver::GenericSolver, node::AbstractRuleNode, path::Vector{Int})

Notify all grammar constraints about the new node and its (grand)children

HerbConstraints.notify_new_nodes — Method

notify_new_nodes(solver::UniformSolver, node::AbstractRuleNode, path::Vector{Int})

Notify all grammar constraints about the new node and its (grand)children

HerbConstraints.notify_tree_manipulation — Method

notify_tree_manipulation(solver::GenericSolver, event_path::Vector{Int})

Notify subscribed constraints that a tree manipulation has occured at the event_path by scheduling them for propagation

HerbConstraints.notify_tree_manipulation — Method

notify_tree_manipulation(solver::UniformSolver, event_path::Vector{Int})

Notify subscribed constraints that a tree manipulation has occured at the event_path by scheduling them for propagation

HerbConstraints.partition — Method

partition(hole::Hole, grammar::ContextSensitiveGrammar)::Vector{BitVector}

Partition a Hole into subdomains grouped by childtypes

HerbConstraints.pattern_match — Method

Generic fallback function for commutativity. Swaps arguments 1 and 2, then dispatches to a more specific signature. If this gets stuck in an infinite loop, the implementation of an AbstractRuleNode type pair is missing.

HerbConstraints.pattern_match — Method

pattern_match(rn::AbstractRuleNode, mn::AbstractRuleNode)::PatternMatchResult

Recursively tries to match AbstractRuleNode rn with AbstractRuleNode mn. Returns a PatternMatchResult that describes if the pattern was matched.

HerbConstraints.pattern_match — Method

pattern_match(node::AbstractRuleNode, domainrulenode::DomainRuleNode, vars::Dict{Symbol, AbstractRuleNode})::PatternMatchResult

Comparing any AbstractRuleNode with a DomainRuleNode

HerbConstraints.pattern_match — Method

pattern_match(rn::AbstractRuleNode, var::VarNode, vars::Dict{Symbol, AbstractRuleNode})::PatternMatchResult

Comparing any AbstractRuleNode with a named VarNode

HerbConstraints.pattern_match — Method

pattern_match(h1::Union{RuleNode, AbstractHole}, h2::Union{RuleNode, AbstractHole}, vars::Dict{Symbol, AbstractRuleNode})::PatternMatchResult

Comparing any pair of Rulenode and/or AbstractHole. It is important to note that some AbstractHoles are already filled and should be treated as RuleNode. This is why this function is dispatched on (isfilled(h1), isfilled(h2)). The '(RuleNode, AbstractHole)' case could still include two nodes of type AbstractHole, but one of them should be treated as a rulenode.

HerbConstraints.pattern_match — Method

pattern_match(rns::Vector{AbstractRuleNode}, mns::Vector{AbstractRuleNode}, vars::Dict{Symbol, AbstractRuleNode})::PatternMatchResult

Pairwise tries to match two ordered lists of AbstractRuleNodes. Typically, this function is used to pattern match the children two AbstractRuleNodes.

HerbConstraints.post! — Method

post!(solver::GenericSolver, constraint::AbstractLocalConstraint)

Imposes the constraint to the current state. By default, the constraint will be scheduled for its initial propagation. Constraints can overload this method to add themselves to notify lists or triggers.

HerbConstraints.post! — Method

post!(solver::UniformSolver, constraint::AbstractLocalConstraint)

Post a new local constraint. Converts the constraint to a state constraint and schedules it for propagation.

HerbConstraints.propagate! — Method

function propagate!(::GenericSolver, ::LocalContainsSubtree)

!!! warning: LocalContainsSubtree uses stateful properties and can therefore not be propagated in the GenericSolver. (The GenericSolver shares constraints among different states, so they cannot use stateful properties)

HerbConstraints.propagate! — Method

function propagate!(solver::Solver, c::LocalContains)

Enforce that the rule appears at or below the path at least once. Uses a helper function to retrieve a list of holes that can potentially hold the target rule. If there is only a single hole that can potentially hold the target rule, that hole will be filled with that rule.

HerbConstraints.propagate! — Method

function propagate!(solver::Solver, c::LocalForbiddenSequence)

HerbConstraints.propagate! — Method

function propagate!(solver::Solver, c::LocalForbidden)

Enforce that the forbidden tree does not occur at the path. The forbidden tree is matched against the AbstractRuleNode located at the path. Deductions are based on the type of the PatternMatchResult returned by the pattern_match function.

HerbConstraints.propagate! — Method

function propagate!(solver::Solver, c::LocalOrdered)

Enforce that the VarNodes in the tree are in the specified order. First the node located at the path is matched to see if the ordered constraint applies here. The nodes matching the variables are stored in the vars dictionary. Then the order is enforced within the make_less_than_or_equal! tree manipulation.

HerbConstraints.propagate! — Method

function propagate!(solver::Solver, c::LocalUnique)

Enforce that the rule appears at or below the path at least once. Uses a helper function to retrieve a list of holes that can potentially hold the target rule. If there is only a single hole that can potentially hold the target rule, that hole will be filled with that rule.

HerbConstraints.propagate! — Method

function propagate!(solver::UniformSolver, c::LocalContainsSubtree)

Enforce that the tree appears at or below the path at least once. Nodes that can potentially become the target sub-tree are considered candidates. In case of multiple candidates, a stateful set of indices is used to keep track of active candidates.

HerbConstraints.remove! — Method

remove!(solver::GenericSolver, path::Vector{Int}, rule_index::Int)

Remove rule_index from the domain of the hole located at the path. It is assumed the path points to a hole, otherwise an exception will be thrown.

HerbConstraints.remove! — Method

remove!(solver::GenericSolver, path::Vector{Int}, rules::Vector{Int})

Remove all rules from the domain of the hole located at the path. It is assumed the path points to a hole, otherwise an exception will be thrown.

HerbConstraints.remove! — Method

remove!(set::StateSparseSet, val::Int)

Removes value val from StateSparseSet set. Returns true if val was in set.

HerbConstraints.remove! — Method

remove!(solver::Solver, path::Vector{Int}, rule_index::Int)

Remove rule_index from the domain of the hole located at the path. It is assumed the path points to a hole, otherwise an exception will be thrown.

HerbConstraints.remove! — Method

remove!(solver::UniformSolver, path::Vector{Int}, rules::Vector{Int})

Remove all rules from the domain of the hole located at the path. It is assumed the path points to a hole, otherwise an exception will be thrown.

HerbConstraints.remove_above! — Method

remove_above!(solver::GenericSolver, path::Vector{Int}, rule_index::Int)

Reduce the domain of the hole located at the path by removing all rules indices above rule_index Example: rule_index = 2. hole with domain [1, 1, 0, 1] gets reduced to [1, 0, 0, 0] and gets simplified to a RuleNode

HerbConstraints.remove_above! — Method

Remove all the values greater than val from the set

HerbConstraints.remove_above! — Method

remove_above!(solver::UniformSolver, path::Vector{Int}, rule_index::Int)

Reduce the domain of the hole located at the path by removing all rules indices above rule_index Example: rule_index = 2. hole with domain {1, 2, 4} gets reduced to {1}

HerbConstraints.remove_all_but! — Method

remove_all_but!(solver::GenericSolver, path::Vector{Int}, new_domain::BitVector)

Reduce the domain of the hole located at the path, to the new_domain. It is assumed the path points to a hole, otherwise an exception will be thrown. It is assumed new_domain ⊆ domain. For example: [1, 0, 1, 0] ⊆ [1, 0, 1, 1]

HerbConstraints.remove_all_but! — Method

remove_all_but!(solver::GenericSolver, path::Vector{Int}, rule_index::Int)

Fill in the hole located at the path with rule rule_index. It is assumed the path points to a hole, otherwise an exception will be thrown. It is assumed rule_index ∈ hole.domain.

!!! warning: If the hole is known to be in the current tree, the hole can be passed directly. The caller has to make sure that the hole instance is actually present at the provided path.

HerbConstraints.remove_all_but! — Method

remove_all_but!(set::StateSparseSet, val::Int)::Bool

Removes all values from StateSparseSet set, except val

HerbConstraints.remove_all_but! — Method

remove_all_but!(solver::UniformSolver, path::Vector{Int}, rule_index::Int)

Fill in the hole located at the path with rule rule_index.

HerbConstraints.remove_below! — Method

remove_below!(solver::GenericSolver, path::Vector{Int}, rule_index::Int)

Reduce the domain of the hole located at the path by removing all rules indices below rule_index Example: rule_index = 2. hole with domain [1, 1, 0, 1] gets reduced to [0, 1, 0, 1]

HerbConstraints.remove_below! — Method

Remove all the values less than val from the set

HerbConstraints.remove_below! — Method

remove_below!(solver::UniformSolver, path::Vector{Int}, rule_index::Int)

Reduce the domain of the hole located at the path by removing all rules indices below rule_index Example: rule_index = 2. hole with domain {1, 2, 4} gets reduced to {2, 4}

HerbConstraints.remove_node! — Method

function remove_node!(solver::GenericSolver, path::Vector{Int})

Remove the node at the given path by substituting it with a hole of the same symbol.

HerbConstraints.restore! — Method

Restores the StateInt stored in the StateIntBackup to its original value

HerbConstraints.restore! — Method

Reverts all the backups since the last save_state!.

HerbConstraints.restore! — Method

Restore state of the solver until the last save_state!

HerbConstraints.save_state! — Method

save_state!(solver::GenericSolver)

Returns a copy of the current state that can be restored by calling load_state!(solver, state)

HerbConstraints.save_state! — Method

Make a backup of the current state. Return to this state by calling restore!.

HerbConstraints.save_state! — Method

Save the current state of the solver, can restored using restore!

HerbConstraints.schedule! — Method

schedule(solver::GenericSolver, constraint::AbstractLocalConstraint)

Schedules the constraint for propagation.

HerbConstraints.set_infeasible! — Method

set_infeasible!(solver::GenericSolver)

Function to be called if any inconsistency has been detected

HerbConstraints.set_infeasible! — Method

set_infeasible!(solver::Solver)

Function to be called if any inconsistency has been detected

HerbConstraints.set_value! — Method

Set the value of the integer to the given val

HerbConstraints.shouldschedule — Method

shouldschedule(solver::Solver, constraint::AbstractLocalConstraint, path::Vector{Int})::Bool

Function that is called when a tree manipulation occured at the path. Returns true if the constraint should be scheduled for propagation.

Default behavior: return true iff the manipulation happened at or below the constraint path.

HerbConstraints.shouldschedule — Method

shouldschedule(::Solver, constraint::LocalForbiddenSequence, path::Vector{Int})::Bool

Return true iff the manipulation happened at or above the constraint path.

HerbConstraints.simplify_hole! — Method

simplify_hole!(solver::GenericSolver, path::Vector{Int})

Takes a Hole and tries to simplify it to a UniformHole or RuleNode. If the domain of the hole is empty, the state will be marked as infeasible

HerbConstraints.substitute! — Method

substitute!(solver::GenericSolver, path::Vector{Int}, new_node::AbstractRuleNode; is_domain_increasing::Union{Nothing, Bool}=nothing)

Substitute the node at the path, with a new_node.

is_domain_increasing: indicates if all grammar constraints should be repropagated from the ground up.

Domain increasing substitutions are substitutions that cannot be achieved by repeatedly removing values from domains. Example of an domain increasing event: hole[{3, 4, 5}] -> hole[{1, 2}]. Example of an domain decreasing event: hole[{3, 4, 5}] -> rulenode(4, [hole[{1, 2}], rulenode(1)]).

HerbCore.contains_hole — Method

contains_hole(hole::StateHole)::Bool

Returns true if the hole or any of its (grand)children are not filled.

HerbCore.get_node_at_location — Method

HerbCore.get_node_at_location(solver::GenericSolver, location::Vector{Int})::AbstractRuleNode

Get the node at path location.

HerbCore.get_node_at_location — Method

get_node_at_location(solver::UniformSolver, path::Vector{Int})

Get the node that is located at the provided path.

HerbCore.get_path — Method

get_path(solver::GenericSolver, node::AbstractRuleNode)

Get the path at which the node is located.

HerbCore.get_path — Method

get_path(solver::UniformSolver, node::AbstractRuleNode)

Get the path at which the node is located.

HerbCore.get_rule — Method

get_rule(hole::StateHole)::Int

Assuming the hole has domain size 1, get the rule it is currently assigned to.

HerbCore.isfilled — Method

isfilled(hole::StateHole)::Bool

Holes with domain size 1 are fixed to a rule. Returns whether the hole has domain size 1. (holes with an empty domain are not considered to be fixed)

Index

HerbConstraints.AbstractGrammarConstraint

HerbConstraints.AbstractLocalConstraint

HerbConstraints.AbstractStateManager

HerbConstraints.Contains

HerbConstraints.ContainsSubtree

HerbConstraints.DomainRuleNode

HerbConstraints.Forbidden

HerbConstraints.ForbiddenSequence

HerbConstraints.GenericSolver

HerbConstraints.GenericSolver

HerbConstraints.GenericSolver

HerbConstraints.LessThanOrEqualHardFail

HerbConstraints.LessThanOrEqualResult

HerbConstraints.LessThanOrEqualSoftFail

HerbConstraints.LessThanOrEqualSuccess

HerbConstraints.LessThanOrEqualSuccessEquality

HerbConstraints.LessThanOrEqualSuccessLessThan

HerbConstraints.LessThanOrEqualSuccessWithHoles

HerbConstraints.LocalContains

HerbConstraints.LocalContainsSubtree

HerbConstraints.LocalContainsSubtree

HerbConstraints.LocalForbidden

HerbConstraints.LocalForbiddenSequence

HerbConstraints.LocalOrdered

HerbConstraints.LocalUnique

HerbConstraints.MakeEqualHardFail

HerbConstraints.MakeEqualResult

HerbConstraints.MakeEqualSoftFail

HerbConstraints.MakeEqualSuccess

HerbConstraints.Ordered

HerbConstraints.PatternMatchHardFail

HerbConstraints.PatternMatchResult

HerbConstraints.PatternMatchSoftFail

HerbConstraints.PatternMatchSuccess

HerbConstraints.PatternMatchSuccessWhenHoleAssignedTo

HerbConstraints.Solver

HerbConstraints.SolverState

HerbConstraints.SolverStatistics

HerbConstraints.StateHole

HerbConstraints.StateHole

HerbConstraints.StateHole

HerbConstraints.StateInt

HerbConstraints.StateIntBackup

HerbConstraints.StateManager

HerbConstraints.StateSparseSet

HerbConstraints.StateSparseSet

HerbConstraints.StateStack

HerbConstraints.StateStack

HerbConstraints.StateStack

HerbConstraints.UniformSolver

HerbConstraints.UniformSolver

HerbConstraints.Unique

HerbConstraints.VarNode

HerbCore.AbstractConstraint

HerbCore.AbstractGrammar

HerbCore.AbstractHole

HerbCore.AbstractRuleNode

HerbCore.AbstractUniformHole

HerbCore.Hole

HerbCore.HoleReference

HerbCore.RuleNode

HerbCore.RuleNode

HerbCore.RuleNode

HerbCore.UniformHole

HerbGrammar.ContextSensitiveGrammar

HerbGrammar.NodeLoc

HerbGrammar.SymbolTable

HerbSearch.AbstractBFSIterator

HerbSearch.AbstractDFSIterator

HerbSearch.AbstractMHSearchIterator

HerbSearch.AbstractSASearchIterator

HerbSearch.AbstractVLSNSearchIterator

HerbSearch.BFSIterator

HerbSearch.DFSIterator

HerbSearch.ExpandFailureReason

HerbSearch.GeneticSearchIterator

HerbSearch.MHSearchIterator

HerbSearch.MLFSIterator

HerbSearch.ProgramIterator

HerbSearch.RandomIterator

HerbSearch.SASearchIterator

HerbSearch.StochasticSearchIterator

HerbSearch.SynthResult

HerbSearch.TopDownIterator

HerbSearch.UniformIterator

HerbSearch.UniformIterator

HerbSearch.VLSNSearchIterator

HerbSpecification.AbstractDependentTypeSpecification

HerbSpecification.AgdaSpecification

HerbSpecification.IOExample

HerbSpecification.MetricProblem

HerbSpecification.Problem

HerbSpecification.SMTSpecification

HerbSpecification.Trace

Base.collect

Base.collect

Base.findall

Base.findfirst

Base.findlast

Base.get

Base.getindex

Base.getindex

Base.in

Base.in

Base.insert!

Base.isless

Base.iterate

Base.iterate

Base.iterate

Base.iterate

Base.iterate

Base.length

Base.length

Base.length

Base.length

Base.push!

Base.rand

Base.rand

Base.rand

Base.show

Base.size

Base.size

Base.sum

HerbConstraints._contains

HerbConstraints._count_occurrences

HerbConstraints._count_occurrences

HerbConstraints._count_occurrences!

HerbConstraints._exchange_positions!

HerbConstraints._update_bounds_val_removed!

HerbConstraints._update_max_val_removed!

HerbConstraints._update_min_val_removed!

HerbConstraints.annotation2constraint

HerbConstraints.are_disjoint

HerbConstraints.are_disjoint

HerbConstraints.backup!

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.contains_varnode

HerbConstraints.deactivate!

HerbConstraints.deactivate!

HerbConstraints.decrement!

HerbConstraints.fix_point!

HerbConstraints.freeze_state

HerbConstraints.get_grammar

HerbConstraints.get_grammar

HerbConstraints.get_hole_at_location

HerbConstraints.get_hole_at_location

HerbConstraints.get_intersection

HerbConstraints.get_max_depth

HerbConstraints.get_max_size

HerbConstraints.get_nodes

HerbConstraints.get_nodes_on_path

HerbConstraints.get_priority

HerbConstraints.get_starting_symbol

HerbConstraints.get_state

HerbConstraints.get_tree

HerbConstraints.get_tree

HerbConstraints.get_tree_size

HerbConstraints.get_value

HerbConstraints.increment!

HerbConstraints.is_subdomain

HerbConstraints.is_subdomain

HerbConstraints.isfeasible

HerbConstraints.isfeasible

HerbConstraints.load_state!

HerbConstraints.make_equal!

HerbConstraints.make_less_than_or_equal!

HerbConstraints.make_less_than_or_equal!

HerbConstraints.make_less_than_or_equal!

HerbConstraints.new_state!

HerbConstraints.notify_new_node

HerbConstraints.notify_new_nodes

HerbConstraints.notify_new_nodes

HerbConstraints.notify_tree_manipulation

HerbConstraints.notify_tree_manipulation

HerbConstraints.partition

HerbConstraints.pattern_match

HerbConstraints.pattern_match

HerbConstraints.pattern_match

HerbConstraints.pattern_match

HerbConstraints.pattern_match

HerbConstraints.pattern_match

HerbConstraints.post!

HerbConstraints.post!

HerbConstraints.propagate!

HerbConstraints.propagate!

HerbConstraints.propagate!

HerbConstraints.propagate!

HerbConstraints.propagate!

HerbConstraints.propagate!

HerbConstraints.propagate!

HerbConstraints.remove!

HerbConstraints.remove!

HerbConstraints.remove!

HerbConstraints.remove!

HerbConstraints.remove!

HerbConstraints.remove_above!

HerbConstraints.remove_above!

HerbConstraints.remove_above!

HerbConstraints.remove_all_but!

HerbConstraints.remove_all_but!

HerbConstraints.remove_all_but!

HerbConstraints.remove_all_but!

HerbConstraints.remove_below!

HerbConstraints.remove_below!

HerbConstraints.remove_below!

HerbConstraints.remove_node!

HerbConstraints.restore!

HerbConstraints.restore!

HerbConstraints.restore!

HerbConstraints.save_state!

HerbConstraints.save_state!

HerbConstraints.save_state!

HerbConstraints.schedule!

HerbConstraints.set_infeasible!

HerbConstraints.set_infeasible!

HerbConstraints.set_value!

HerbConstraints.shouldschedule

HerbConstraints.shouldschedule

HerbConstraints.simplify_hole!

HerbConstraints.substitute!

HerbCore.contains_hole

HerbCore.contains_hole

HerbCore.contains_index

HerbCore.contains_nonuniform_hole

HerbCore.depth

HerbCore.get_children

HerbCore.get_node_at_location

HerbCore.get_node_at_location

HerbCore.get_node_at_location

HerbCore.get_node_at_location

HerbCore.get_path

HerbCore.get_path

HerbCore.get_path

HerbCore.get_rule

HerbCore.get_rule

HerbCore.get_rulesequence

HerbCore.hasdynamicvalue

HerbCore.have_same_shape

HerbCore.isfilled

HerbCore.isfilled

HerbCore.isuniform

HerbCore.node_depth

HerbCore.number_of_holes

HerbCore.rulesoftype

HerbCore.rulesoftype

HerbCore.rulesonleft

HerbCore.swap_node

HerbCore.swap_node

HerbGrammar.add_rule!

HerbGrammar.add_rule!

HerbGrammar.add_rule!

HerbGrammar.addconstraint!

HerbGrammar.child_types

HerbGrammar.child_types

HerbGrammar.cleanup_removed_rules!

HerbGrammar.clearconstraints!

HerbGrammar.containedin

HerbGrammar.contains_returntype

HerbGrammar.expr2csgrammar

HerbGrammar.expr2pcsgrammar

HerbGrammar.expr2rulenode

HerbGrammar.expr2rulenode

HerbGrammar.expr2rulenode

HerbGrammar.expr2rulenode

HerbGrammar.get_childtypes

HerbGrammar.get_domain

HerbGrammar.get_domain

HerbGrammar.get_rulesequence

HerbGrammar.grammar2symboltable

HerbGrammar.init_probabilities!

HerbGrammar.iscomplete

HerbGrammar.iseval

HerbGrammar.iseval

HerbGrammar.iseval

HerbGrammar.isprobabilistic

HerbGrammar.isterminal

HerbGrammar.isterminal

HerbGrammar.isterminal

HerbGrammar.isvariable

HerbGrammar.isvariable

HerbGrammar.isvariable

HerbGrammar.isvariable

HerbGrammar.log_probability

HerbGrammar.max_arity

HerbGrammar.max_rulenode_log_probability

HerbGrammar.merge_grammars!

HerbGrammar.mindepth

HerbGrammar.mindepth_map

HerbGrammar.nchildren

HerbGrammar.nchildren

HerbGrammar.nonterminals

HerbGrammar.normalize!

HerbGrammar.parse_probabilistic_rule

HerbGrammar.probability

HerbGrammar.read_csg

HerbGrammar.read_pcsg

HerbGrammar.remove_rule!

HerbGrammar.return_type

HerbGrammar.return_type

HerbGrammar.return_type

HerbGrammar.root_node_loc

HerbGrammar.rulenode2expr

HerbGrammar.rulenode_log_probability

HerbGrammar.rulesonleft

HerbGrammar.store_csg

HerbGrammar.subsequenceof

HerbGrammar.swap_node

HerbGrammar.swap_node

HerbInterpret.evaluate_program

HerbInterpret.execute_on_input

HerbInterpret.execute_on_input

HerbInterpret.execute_on_input

HerbInterpret.execute_on_input

HerbInterpret.interpret

HerbInterpret.test_all_examples

HerbInterpret.test_examples

HerbSearch._calculate_cost

HerbSearch._find_next_complete_tree

HerbSearch.accept

HerbSearch.best_accept

HerbSearch.calculate_cost

HerbSearch.const_temperature

HerbSearch.constructNeighbourhood

HerbSearch.constructNeighbourhoodRuleSubset

HerbSearch.cross_over

HerbSearch.crossover_swap_children_1

HerbSearch.crossover_swap_children_2

HerbSearch.decreasing_temperature

HerbSearch.default_fitness

HerbSearch.derivation_heuristic

HerbSearch.derivation_heuristic

HerbSearch.enumerate_neighbours_propose

HerbSearch.evaluate

HerbSearch.extract_name_from_argument

HerbSearch.fitness

HerbSearch.generate_branches

HerbSearch.get_best_program

HerbSearch.heuristic_leftmost

HerbSearch.heuristic_random

HerbSearch.heuristic_rightmost

HerbSearch.heuristic_smallest_domain

HerbSearch.hole_heuristic

HerbSearch.is_field_decl

HerbSearch.is_kwdef

HerbSearch.mean_squared_error

HerbSearch.misclassification

HerbSearch.mutate!

HerbSearch.mutate_random!

HerbSearch.neighbourhood

HerbSearch.next_solution!

HerbSearch.priority_function

HerbSearch.priority_function

HerbSearch.priority_function

HerbSearch.priority_function

HerbSearch.probabilistic_accept

HerbSearch.probabilistic_accept_with_temperature

HerbSearch.probabilistic_accept_with_temperature_fraction

HerbSearch.processkwarg!

HerbSearch.propose

HerbSearch.random_fill_propose

HerbSearch.select_chromosome

HerbSearch.select_fitness_proportional_parents

HerbSearch.select_parents

HerbSearch.set_stateholes!

HerbSearch.synth

HerbSearch.temperature

HerbSearch.validate_iterator

StatsBase.sample

StatsBase.sample

StatsBase.sample

StatsBase.sample

HerbConstraints.@csgrammar_annotated

HerbCore.@rulenode

HerbGrammar.@cfgrammar

HerbGrammar.@csgrammar

HerbGrammar.@pcsgrammar

HerbSearch.@programiterator