HerbCore.jl Documentation

HerbCore.AbstractRuleNode — Type

AbstractRuleNode

Abstract type for representing expression trees. Expression trees consist of RuleNodes and Holes.

A RuleNode represents a certain production rule in the Grammar.

A Hole is a placeholder where certain rules in the grammar still can be applied.

HerbCore.Constraint — Type

Represents a constraint for a ContextSensitiveGrammar. Concrete implementations can be found in HerbConstraints.jl.

HerbCore.Grammar — Type

Grammar

Abstract type representing all grammars. It is assumed that all grammar structs have at least the following attributes:

rules::Vector{Any}: A list of RHS of rules (subexpressions).

types::Vector{Symbol}: A list of LHS of rules (types, all symbols).

isterminal::BitVector: A bitvector where bit i represents whether rule i is terminal.

iseval::BitVector: A bitvector where bit i represents whether rule i is an eval rule.

bytype::Dict{Symbol,Vector{Int}}: A dictionary that maps a type to all rules of said type.

domains::Dict{Symbol, BitVector}: A dictionary that maps a type to a domain bitvector. The domain bitvector has bit i set to true iff the ith rule is of this type.

childtypes::Vector{Vector{Symbol}}: A list of types of the children for each rule.

If a rule is terminal, the corresponding list is empty.

log_probabilities::Union{Vector{Real}, Nothing}: A list of probabilities for each rule.

If the grammar is non-probabilistic, the list can be nothing.

For concrete types, see ContextFreeGrammar and ContextSensitiveGrammar.

HerbCore.Hole — Type

Hole <: AbstractRuleNode

A Hole is a placeholder where certain rules from the grammar can still be applied. The domain of a Hole defines which rules can be applied. The domain is a bitvector, where the ith bit is set to true if the ith rule in the grammar can be applied.

HerbCore.HoleReference — Type

HoleReference

Contains a hole and the path to the hole from the root of the tree.

HerbCore.RuleNode — Type

RuleNode <: AbstractRuleNode

A RuleNode represents a node in an expression tree. Each node corresponds to a certain rule in the Grammar. A RuleNode consists of:

ind: The index of the rule in the Grammar which this node is representing.

_val: Field for storing immediately evaluated values

children: The children of this node in the expression tree

Compat

Evaluate immediately functionality is not yet supported by most of Herb.jl.

HerbCore.RuleNode — Method

RuleNode(ind::Int, _val::Any)

Create a RuleNode for the Grammar rule with index ind, _val as immediately evaluated value and no children

Warning

Only use this constructor if you are absolutely certain that a rule is terminal and cannot have children. Use [RuleNode(ind::Int, grammar::Grammar)] for rules that might have children. In general, Holes should be used as a placeholder when the children of a node are not yet known.

Compat

Evaluate immediately functionality is not yet supported by most of Herb.jl.

HerbCore.RuleNode — Method

RuleNode(ind::Int, children::Vector{AbstractRuleNode})

Create a RuleNode for the Grammar rule with index ind and children as subtrees.

Base.isless — Method

Base.isless(rn₁::AbstractRuleNode, rn₂::AbstractRuleNode)::Bool

Compares two RuleNodes. Returns true if the left RuleNode is less than the right RuleNode. Order is determined from the index of the RuleNodes. If both RuleNodes have the same index, a depth-first search is performed in both RuleNodes until nodes with a different index are found.

Base.length — Method

Base.length(root::RuleNode)

Return the number of nodes in the tree rooted at root. Holes don't count.

Base.length — Method

Base.length(root::RuleNode)

Return the number of nodes in the tree rooted at root. Holes don't count.

HerbCore.contains_hole — Method

contains_hole(rn::RuleNode) = any(contains_hole(c) for c ∈ rn.children)

Checks if an AbstractRuleNode tree contains a Hole.

HerbCore.depth — Method

depth(root::RuleNode)::Int

Return the depth of the AbstractRuleNode tree rooted at root. Holes don't count towards the depth.

HerbCore.get_node_at_location — Method

get_node_at_location(root::RuleNode, location::Vector{Int})

Retrieves a RuleNode at the given location by reference.

HerbCore.get_rulesequence — Method

get_rulesequence(node::RuleNode, path::Vector{Int})

Extract the derivation sequence from a path (sequence of child indices) and an AbstractRuleNode. If the path is deeper than the deepest node, it returns what it has.

HerbCore.node_depth — Method

node_depth(root::AbstractRuleNode, node::AbstractRuleNode)::Int

Return the depth of node for an AbstractRuleNode tree rooted at root. Depth is 1 when root == node.

Warning

node must be a subtree of root in order for this function to work.

HerbCore.rulesoftype — Method

rulesoftype(node::RuleNode, ruleset::Set{Int})

Returns every rule in the ruleset that is also used in the AbstractRuleNode tree.

HerbCore.rulesonleft — Method

rulesonleft(expr::RuleNode, path::Vector{Int})::Set{Int}

Finds all rules that are used in the left subtree defined by the path.

HerbCore.swap_node — Method

swap_node(expr::AbstractRuleNode, new_expr::AbstractRuleNode, path::Vector{Int})

Replace a node in expr, specified by path, with new_expr. Path is a sequence of child indices, starting from the root node.

HerbCore.swap_node — Method

swap_node(expr::RuleNode, node::RuleNode, child_index::Int, new_expr::RuleNode)

Replace child i of a node, a part of larger expr, with new_expr.

Index

HerbConstraints.AbstractMatchNode

HerbConstraints.ComesAfter

HerbConstraints.ComesAfter

HerbConstraints.Forbidden

HerbConstraints.ForbiddenPath

HerbConstraints.GrammarContext

HerbConstraints.LocalConstraint

HerbConstraints.LocalForbidden

HerbConstraints.LocalOneOf

HerbConstraints.LocalOrdered

HerbConstraints.MatchFail

HerbConstraints.MatchNode

HerbConstraints.MatchVar

HerbConstraints.OneOf

HerbConstraints.Ordered

HerbConstraints.PropagatorConstraint

HerbConstraints.RequireOnLeft

HerbCore.AbstractRuleNode

HerbCore.Constraint

HerbCore.Grammar

HerbCore.Hole

HerbCore.HoleReference

HerbCore.RuleNode

HerbCore.RuleNode

HerbCore.RuleNode

HerbData.IOExample

HerbData.IOPExample

HerbData.Problem

HerbGrammar.ContextFreeGrammar

HerbGrammar.ContextSensitiveGrammar

HerbGrammar.NodeLoc

HerbGrammar.SymbolTable

HerbGrammar.SymbolTable

HerbSearch.ContextSensitivePriorityEnumerator

HerbSearch.ExpandFailureReason

HerbSearch.ExpressionIterator

HerbSearch.GeneticSearchIterator

HerbSearch.PriorityQueueItem

HerbSearch.PriorityQueueItem

HerbSearch.PropagateResult

HerbSearch.StochasticSearchEnumerator

Base.get

Base.insert!

Base.isless

Base.iterate

Base.iterate

Base.iterate

Base.iterate

Base.iterate

Base.length

Base.length

Base.rand

Base.rand

Base.rand

Base.show

Base.show

HerbConstraints._matchnode2expr

HerbConstraints._matchnode2expr

HerbConstraints._matchnode2expr

HerbConstraints._matchnode2expr

HerbConstraints._pattern_match

HerbConstraints._pattern_match

HerbConstraints._pattern_match_with_hole

HerbConstraints._pattern_match_with_hole

HerbConstraints._pattern_match_with_hole

HerbConstraints._pattern_match_with_hole

HerbConstraints._rulenode_compare

HerbConstraints._rulenode_match_with_hole

HerbConstraints.addparent!

HerbConstraints.annotation2constraint

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.copy_and_insert

HerbConstraints.make_smaller_or_equal

HerbConstraints.matchnode2expr

HerbConstraints.matchnode2expr

HerbConstraints.propagate

HerbConstraints.propagate

HerbConstraints.propagate

HerbConstraints.propagate

HerbConstraints.propagate

HerbConstraints.propagate

HerbConstraints.propagate

HerbConstraints.propagate

HerbCore.contains_hole

HerbCore.depth

HerbCore.get_node_at_location

HerbCore.get_rulesequence

HerbCore.node_depth

HerbCore.rulesoftype

HerbCore.rulesonleft

HerbCore.swap_node

HerbCore.swap_node

HerbData.read_IOPexamples

HerbData.read_IOexamples

HerbData.readdata

HerbData.readfile

HerbData.write_IOPexamples

HerbData.write_IOexamples

HerbEvaluation.evaluate_program

HerbEvaluation.execute_on_examples

HerbEvaluation.interpret

HerbEvaluation.test_all_examples

HerbEvaluation.test_examples

HerbEvaluation.test_with_input

HerbGrammar.add_rule!

HerbGrammar.addconstraint!

HerbGrammar.cfg2csg

HerbGrammar.child_types

HerbGrammar.child_types

HerbGrammar.cleanup_removed_rules!

HerbGrammar.clearconstraints!

HerbGrammar.containedin

HerbGrammar.contains_returntype

HerbGrammar.expr2cfgrammar

HerbGrammar.expr2csgrammar

HerbGrammar.expr2pcfgrammar

HerbGrammar.expr2pcsgrammar

HerbGrammar.get_childtypes

HerbGrammar.get_domain

HerbGrammar.get_domain

HerbGrammar.get_node_at_location

HerbGrammar.get_rulesequence

HerbGrammar.has_children

HerbGrammar.iscomplete

HerbGrammar.iseval

HerbGrammar.iseval

HerbGrammar.iseval

HerbGrammar.isprobabilistic

HerbGrammar.isterminal

HerbGrammar.isterminal

HerbGrammar.isterminal

HerbGrammar.isvariable

HerbGrammar.log_probability

HerbGrammar.max_arity

HerbGrammar.mindepth

HerbGrammar.mindepth_map

HerbGrammar.nchildren

HerbGrammar.nchildren

HerbGrammar.nonterminals

HerbGrammar.probability

HerbGrammar.read_cfg

HerbGrammar.read_csg

HerbGrammar.read_pcfg

HerbGrammar.read_pcsg

HerbGrammar.remove_rule!

HerbGrammar.return_type

HerbGrammar.return_type

HerbGrammar.root_node_loc

HerbGrammar.rulenode2expr

HerbGrammar.rulenode_log_probability

HerbGrammar.rulesoftype

HerbGrammar.rulesoftype

HerbGrammar.rulesoftype

HerbGrammar.rulesonleft

HerbGrammar.store_cfg

HerbGrammar.store_csg

HerbGrammar.subsequenceof

HerbGrammar.swap_node

HerbGrammar.swap_node

HerbSearch._expand

HerbSearch._expand

HerbSearch._find_next_complete_tree

HerbSearch.best_accept

HerbSearch.calculate_cost

HerbSearch.const_temperature

HerbSearch.constructNeighbourhood

HerbSearch.constructNeighbourhoodRuleSubset

HerbSearch.count_expressions

HerbSearch.count_expressions

HerbSearch.crossover_swap_children_1

HerbSearch.crossover_swap_children_2

HerbSearch.decreasing_temperature

HerbSearch.default_error_function

HerbSearch.default_fitness

HerbSearch.enumerate_neighbours_propose

HerbSearch.get_best_program

HerbSearch.get_bfs_enumerator

HerbSearch.get_bfs_enumerator

HerbSearch.get_dfs_enumerator

HerbSearch.get_dfs_enumerator

HerbSearch.get_genetic_enumerator

HerbSearch.get_mh_enumerator

HerbSearch.get_most_likely_first_enumerator

HerbSearch.get_most_likely_first_enumerator

HerbSearch.get_sa_enumerator

HerbSearch.get_vlsn_enumerator

HerbSearch.heuristic_leftmost

HerbSearch.heuristic_random

HerbSearch.heuristic_rightmost

HerbSearch.heuristic_smallest_domain

HerbSearch.mean_squared_error

HerbSearch.misclassification

HerbSearch.most_likely_priority_function

HerbSearch.mse_error_function

HerbSearch.mutate_random!

HerbSearch.probabilistic_accept

HerbSearch.probabilistic_accept_with_temperature

HerbSearch.probabilistic_accept_with_temperature_fraction

HerbSearch.propagate_constraints

HerbSearch.random_fill_propose

HerbSearch.search

HerbSearch.search_best

HerbSearch.search_rulenode

HerbSearch.select_chromosome

HerbSearch.select_fitness_proportional_parents

HerbSearch.validate_iterator

StatsBase.sample

StatsBase.sample

StatsBase.sample

StatsBase.sample

HerbConstraints.@csgrammar_annotated

HerbGrammar.@cfgrammar

HerbGrammar.@csgrammar

HerbGrammar.@pcfgrammar

HerbGrammar.@pcsgrammar