HerbSearch.jl Documentation

HerbSearch.ContextSensitivePriorityEnumerator — Type

mutable struct ContextSensitivePriorityEnumerator <: ExpressionIterator

Enumerates a context-free grammar starting at Symbol sym with respect to the grammar up to a given depth and a given size. The exploration is done using the given priority function for derivations, and the expand function for discovered nodes.

HerbSearch.ExpandFailureReason — Type

@enum ExpandFailureReason limit_reached=1 already_complete=2

Representation of the different reasons why expanding a partial tree failed. Currently, there are two possible causes of the expansion failing:

limit_reached: The depth limit or the size limit of the partial tree would be violated by the expansion

already_complete: There is no hole left in the tree, so nothing can be expanded.

HerbSearch.ExpressionIterator — Type

abstract type ExpressionIterator

Abstract super-type for all possible enumerators.

HerbSearch.GeneticSearchIterator — Type

GeneticSearchIterator{FitnessFunction,CrossOverFunction,MutationFunction,SelectParentsFunction,EvaluationFunction} <: ExpressionIterator

Defines an ExpressionIterator using genetic search.

Consists of:

grammar::ContextSensitiveGrammar: the grammar to search over

examples::Vector{<:Example}: a collection of examples defining the specification

fitness::FitnessFunction: assigns a numerical value (fitness score) to each individual based on how closely it meets the desired objective

cross_over::CrossOverFunction: combines the program from two parent individuals to create one or more offspring individuals

mutation!::MutationFunction: mutates the program of an invididual

select_parents::SelectParentsFunction: selects two parents for the crossover

evaluation_function::EvaluationFunction: interpreter to evaluate the individual programs

start_symbol::Symbol: defines the start symbol from which the search should be started

population_size::Int64: number of inviduals in the population

mutation_probability::Float64: probability of mutation for each individual

maximum_initial_population_depth::Int64: maximum depth of trees when population is initialized

end

HerbSearch.PriorityQueueItem — Type

struct PriorityQueueItem

Represents an item in the priority enumerator priority queue. An item contains of:

tree: A partial AST

size: The size of the tree. This is a cached value which prevents having to traverse the entire tree each time the size is needed.

constraints: The local constraints that apply to this tree. These constraints are enforced each time the tree is modified.

HerbSearch.PriorityQueueItem — Method

PriorityQueueItem(tree::AbstractRuleNode, size::Int)

Constructs PriorityQueueItem given only a tree and the size, but no constraints.

HerbSearch.PropagateResult — Type

@enum PropagateResult tree_complete=1 tree_incomplete=2 tree_infeasible=3

Representation of the possible results of a constraint propagation. At the moment there are three possible outcomes:

tree_complete: The propagation was applied successfully and the tree does not contain any holes anymore. Thus no constraints can be applied anymore.

tree_incomplete: The propagation was applied successfully and the tree does contain more holes. Thus more constraints may be applied to further prune the respective domains.

tree_infeasible: The propagation was succesful, but there are holes with empty domains. Hence, the tree is now infeasible.

HerbSearch.StochasticSearchEnumerator — Type

Base.@kwdef struct StochasticSearchEnumerator <: ExpressionIterator

A unified struct for the algorithms Metropolis Hastings, Very Large Scale Neighbourhood and Simulated Annealing. Each algorithm implements neighbourhood propose accept and temperature functions. Below the signiture of all this function is shown

Signatures

Returns a node location from the program that is the neighbourhood. It can also return other information using dict

neighbourhood(program::RuleNode, grammar::Grammar) -> (loc::NodeLocation, dict::Dict)

Proposes a list of programs using the location provided by neighbourhood and the dict.

propose(current_program, loc::NodeLocation, grammar::Grammar, max_depth::Int64, dict::Dict) -> Iter[RuleNode]

Based on the current program and possible cost and temperature it accepts the program or not. Usually we would always want to accept better programs but we might get stuck if we do so. That is why some implementations of the accept function accept with a probability costs that are worse. cost means how different are the outcomes of the program compared to the correct outcomes. The lower the cost the better the program performs on the examples. The cost is provided by the cost_function

accept(current_cost::Real, possible_cost::Real, temperature::Real) -> Bool

Returns the new temperature based on the previous temperature. Higher the temperature means that the algorithm will explore more.

temperature(previous_temperature::Real) -> Real

Returns the cost of the current program. It receives a list of tuples (expected, found) and gives back a cost.

cost_function(outcomes::Tuple{<:Number,<:Number}[]) -> Real

Fields

grammar::ContextSensitiveGrammar grammar that the algorithm uses

max_depth::Int64 = 5 maximum depth of the program to generate

examples::Vector{Example} example used to check the program

neighbourhood::Function

propose::Function

accept::Function

temperature::Function

cost_function::Function

start_symbol::Symbol the start symbol of the algorithm :Real or :Int

initial_temperature::Real = 1

evaluation_function::Function that evaluates the julia expressions

An iterator over all possible expressions of a grammar up to max_depth with start symbol sym.

Base.iterate — Method

Base.iterate(iter::ContextSensitivePriorityEnumerator, pq::DataStructures.PriorityQueue)

Describes the iteration for a given ContextSensitivePriorityEnumerator and a PriorityQueue over the grammar without enqueueing new items to the priority queue. Recursively returns the result for the priority queue.

Base.iterate — Method

Base.iterate(iter::ContextSensitivePriorityEnumerator)

Describes the iteration for a given ContextSensitivePriorityEnumerator over the grammar. The iteration constructs a PriorityQueue first and then prunes it propagating the active constraints. Recursively returns the result for the priority queue.

Base.iterate — Method

Base.iterate(iter::GeneticSearchIterator, current_state::GeneticIteratorState)

Iterates the search space using a genetic algorithm. Takes the iterator and the current state to mutate and crossover random inviduals. Returns the best program-so-far and the state of the iterator.

Base.iterate — Method

Base.iterate(iter::GeneticSearchIterator)

Iterates the search space using a genetic algorithm. First generates a population sampling random programs. Returns the best program-so-far, and the state of the iterator.

Base.iterate — Method

Base.iterate(iter::StochasticSearchEnumerator, current_state::IteratorState)

The algorithm that constructs the iterator of StochasticSearchEnumerator. It has the following structure:

get a random node location -> location,dict = neighbourhood(current_program)

call propose on the current program getting a list of possbile replacements in the node location

iterate through all the possible replacements and perform the replacement in the current program

accept the new program by modifying the next_program or reject the new program

return the new next_program

Base.rand — Function

rand(::Type{RuleNode}, grammar::Grammar, typ::Symbol, max_depth::Int=10)

Generates a random RuleNode of return type typ and maximum depth max_depth.

Base.rand — Function

rand(::Type{RuleNode}, grammar::Grammar, max_depth::Int=10)

Generates a random RuleNode of arbitrary type and maximum depth max_depth.

Base.rand — Function

rand(::Type{RuleNode}, grammar::Grammar, typ::Symbol, dmap::AbstractVector{Int}, max_depth::Int=10)

Generates a random RuleNode, i.e. an expression tree, of root type typ and maximum depth max_depth guided by a depth map dmap if possible.

HerbSearch._expand — Method

_expand(node::Hole, grammar::ContextSensitiveGrammar, ::Int, max_holes::Int, context::GrammarContext, hole_heuristic::Function, derivation_heuristic::Function)::Union{ExpandFailureReason, Vector{TreeConstraints}}

Expands a given hole that was found in _expand using the given derivation heuristic. Returns the list of discovered nodes in that order and with their respective constraints.

HerbSearch._expand — Method

_expand(root::RuleNode, grammar::ContextSensitiveGrammar, max_depth::Int, max_holes::Int, context::GrammarContext, hole_heuristic::Function, derivation_heuristic::Function)::Union{ExpandFailureReason, Vector{TreeConstraints}}

Recursive expand function used in multiple enumeration techniques. Expands one hole/undefined leaf of the given RuleNode tree found using the given hole heuristic. If the expansion was successful, returns a list of new trees and a list of lists of hole locations, corresponding to the holes of each newly expanded tree. Returns nothing if tree is already complete (i.e. contains no holes). Returns an empty list if the tree is partial (i.e. contains holes), but they could not be expanded because of the depth limit.

HerbSearch._find_next_complete_tree — Method

_find_next_complete_tree(grammar::ContextSensitiveGrammar, max_depth::Int, max_size::Int, priority_function::Function, expand_function::Function, pq::PriorityQueue)::Union{Tuple{RuleNode, PriorityQueue}, Nothing}

Takes a priority queue and returns the smallest AST from the grammar it can obtain from the queue or by (repeatedly) expanding trees that are in the queue. Returns nothing if there are no trees left within the depth limit.

HerbSearch.best_accept — Method

best_accept(current_cost::Real, next_cost::Real, temperature)

Returns true if the cost of the proposed program is smaller than the cost of the current program. Otherwise, returns false.

Arguments

current_cost::Real: the cost of the current program.

next_cost::Real: the cost of the proposed program.

temperature::Real: the temperature; not used.

HerbSearch.calculate_cost — Method

calculate_cost(program::RuleNode, cost_function::Function, examples::AbstractVector{Example}, grammar::Grammar, evaluation_function::Function)

Returns the cost of the program using the examples and the cost_function. It first convert the program to an expression and evaluates it on all the examples using HerbEvaluationevaluate_program.

HerbSearch.const_temperature — Method

const_temperature(current_temperature)

Returns the temperature unchanged. This function is used by Metropolis Hastings and Very Large Neighbourhood Search algorithms.

Arguments

current_temperature::Real: the current temperature of the search.

HerbSearch.constructNeighbourhood — Method

constructNeighbourhood(current_program::RuleNode, grammar::Grammar)

The neighbourhood node location is chosen at random. The dictionary is nothing.

Arguments

current_program::RuleNode: the current program.

grammar::Grammar: the grammar.

HerbSearch.constructNeighbourhoodRuleSubset — Method

constructNeighbourhoodRuleSubset(current_program::RuleNode, grammar::Grammar)

The neighbourhood node location is chosen at random. The dictionary is contains one entry with key "rule_subset" and value of type Vector{Any} being a random subset of grammar rules.

Arguments

current_program::RuleNode: the current program.

grammar::Grammar: the grammar.

HerbSearch.count_expressions — Method

count_expressions(iter::ExpressionIterator)

Counts and returns the number of possible expressions in the expression iterator. The Iterator is not modified.

HerbSearch.count_expressions — Method

count_expressions(grammar::Grammar, max_depth::Int, max_size::Int, sym::Symbol)

Counts and returns the number of possible expressions of a grammar up to max_depth with start symbol sym.

HerbSearch.crossover_swap_children_1 — Method

crossover_swap_children_1(parent1::RuleNode, parent2::RuleNode)

Performs a random crossover of two parents of type RuleNode. The subprograms are swapped and only one altered parent program is returned.

HerbSearch.crossover_swap_children_2 — Method

crossover_swap_children_2(parent1::RuleNode, parent2::RuleNode)

Performs a random crossover of two parents of type RuleNode. The subprograms are swapped and both altered parent programs are returned.

HerbSearch.decreasing_temperature — Method

decreasing_temperature(percentage::Real)

Returns a function that produces a temperature decreased by percentage%. This function is used by the Simmulated Annealing algorithm.

Arguments

percentage::Real: the percentage to decrease the temperature by.

HerbSearch.default_error_function — Method

default_error_function(old_error, output, expected_output)

Default error function for search_best.

- old_error         - The existing total error
- output            - The actual output of the evaluator
- expected_output   - The expected output for the example

The default function returns 0 if the outputs match and 1 otherwise.

HerbSearch.default_fitness — Method

default_fitness(program, results)

Defines the default fitness function taking the program and its results. Results are a vector of tuples, where each tuple is in the form Tuple{expected_output, actual_output}. As we are looking for individuals with the highest fitness function, the error is inverted.

HerbSearch.enumerate_neighbours_propose — Method

enumerate_neighbours_propose(enumeration_depth::Int64)

The return function is a function that produces a list with all the subprograms with depth at most enumeration_depth.

Arguments

enumeration_depth::Int64: the maximum enumeration depth.

HerbSearch.get_best_program — Method

get_best_program(population::Array{RuleNode}, iter:: GeneticSearchIterator)::RuleNode

Returns the best program within the population with respect to the fitness function.

HerbSearch.get_bfs_enumerator — Function

get_bfs_enumerator(grammar::ContextSensitiveGrammar, max_depth::Int, max_size::Int, sym::Symbol, hole_heuristic::Function=heuristic_leftmost, derivation_heuristic::Function=(a,_) -> a)::ContextSensitivePriorityEnumerator

Returns a breadth-first search enumerator given a ContextSensitiveGrammar. Returns trees in the grammar in increasing order of size.

HerbSearch.get_bfs_enumerator — Function

get_bfs_enumerator(grammar::ContextFreeGrammar, max_depth::Int, max_size::Int, sym::Symbol, hole_heuristic::Function=heuristic_leftmost, derivation_heuristic::Function=(a,_) -> a)::ContextSensitivePriorityEnumerator

Returns a breadth-first search enumerator given a ContextFreeGrammar. Returns trees in the grammar in increasing order of size.

HerbSearch.get_dfs_enumerator — Function

get_dfs_enumerator(grammar::ContextSensitiveGrammar, max_depth::Int, max_size::Int, sym::Symbol, hole_heuristic::Function=heuristic_leftmost, derivation_heuristic::Function=(a,_) -> a)::ContextSensitivePriorityEnumerator

Returns a depth-first search enumerator given a ContextSensitiveGrammar. Returns trees in the grammar in decreasing order of size.

HerbSearch.get_dfs_enumerator — Function

get_dfs_enumerator(grammar::ContextFreeGrammar, max_depth::Int, max_size::Int, sym::Symbol, hole_heuristic::Function=heuristic_leftmost, derivation_heuristic::Function=(a,_) -> a)::ContextSensitivePriorityEnumerator

Returns a depth-first search enumerator given a ContextFreeGrammar. Returns trees in the grammar in decreasing order of size.

HerbSearch.get_genetic_enumerator — Method

get_genetic_enumerator(examples; fitness_function = HerbSearch.default_fitness, initial_population_size = 10, maximum_initial_population_depth = 3, mutation_probability = 0.1, cross_over = HerbSearch.crossover_swap_children_2, select_parents = HerbSearch.select_fitness_proportional_parents, evaluation_function::Function=HerbEvaluation.test_with_input)

Returns a GeneticSearchIterator given a grammar. The iterator is fitted against the examples provided evaluated using the fitness function. All other arguments are hyperparameters for the genetic search procedure.

HerbSearch.get_mh_enumerator — Function

get_mh_enumerator(examples::AbstractArray{<:Example}, cost_function::Function, evaluation_function::Function=HerbEvaluation.test_with_input)

Returns an enumerator that runs according to the Metropolis Hastings algorithm.

examples : array of examples

cost_function : cost function to evaluate the programs proposed

evaluation_function : evaluation function that evaluates the program generated and produces an output

The propose function is randomfillpropose and the accept function is probabilistic. The temperature value of the algorithm remains constant over time.

HerbSearch.get_most_likely_first_enumerator — Function

get_most_likely_first_enumerator(grammar::ContextSensitiveGrammar, max_depth::Int, max_size::Int, sym::Symbol, hole_heuristic::Function=heuristic_leftmost, derivation_heuristic::Function=(a,_) -> a)::ContextSensitivePriorityEnumerator

Returns an enumerator that enumerates expressions in the grammar in decreasing order of probability. Only use this function with probabilistic grammars.

HerbSearch.get_most_likely_first_enumerator — Function

get_most_likely_first_enumerator(grammar::ContextFreeGrammar, max_depth::Int, max_size::Int, sym::Symbol, hole_heuristic::Function=heuristic_leftmost, derivation_heuristic::Function=(a,_) -> a)::ContextSensitivePriorityEnumerator

Returns an enumerator that enumerates expressions in the grammar in decreasing order of probability. Only use this function with probabilistic grammars.

HerbSearch.get_sa_enumerator — Function

get_sa_enumerator(examples, cost_function, initial_temperature=1, temperature_decreasing_factor = 0.99, evaluation_function::Function=HerbEvaluation.test_with_input)

Returns an enumerator that runs according to the Very Large Scale Neighbourhood Search algorithm.

examples : array of examples

cost_function : cost function to evaluate the programs proposed

initial_temperature : the starting temperature of the algorithm

temperature_decreasing_factor : the decreasing factor of the temperature of the time

evaluation_function : evaluation function that evaluates the program generated and produces an output

The propose function is random_fill_propose (the same as for Metropolis Hastings). The accept function is probabilistic but takes into account the tempeerature too.

HerbSearch.get_vlsn_enumerator — Function

get_vlsn_enumerator(examples, cost_function, enumeration_depth = 2, evaluation_function::Function=HerbEvaluation.test_with_input)

Returns an enumerator that runs according to the Very Large Scale Neighbourhood Search algorithm.

examples : array of examples

cost_function : cost function to evaluate the programs proposed

enumeration_depth : the enumeration depth to search for a best program at a time

evaluation_function : evaluation function that evaluates the program generated and produces an output

The propose function consists of all possible programs of the given enumeration_depth. The accept function accepts the program with the lowest cost according to the cost_function. The temperature value of the algorithm remains constant over time.

HerbSearch.heuristic_leftmost — Method

heuristic_leftmost(node::AbstractRuleNode, max_depth::Int)::Union{ExpandFailureReason, HoleReference}

Defines a heuristic over holes, where the left-most hole always gets considered first. Returns a HoleReference once a hole is found. This is the default option for enumerators.

HerbSearch.heuristic_random — Method

heuristic_random(node::AbstractRuleNode, max_depth::Int)::Union{ExpandFailureReason, HoleReference}

Defines a heuristic over holes, where random holes get chosen randomly using random exploration. Returns a HoleReference once a hole is found.

HerbSearch.heuristic_rightmost — Method

heuristic_rightmost(node::AbstractRuleNode, max_depth::Int)::Union{ExpandFailureReason, HoleReference}

Defines a heuristic over holes, where the right-most hole always gets considered first. Returns a HoleReference once a hole is found.

HerbSearch.heuristic_smallest_domain — Method

heuristic_smallest_domain(node::AbstractRuleNode, max_depth::Int)::Union{ExpandFailureReason, HoleReference}

Defines a heuristic over all available holes in the unfinished AST, by considering the size of their respective domains. A domain here describes the number of possible derivations with respect to the constraints. Returns a HoleReference once a hole is found.

HerbSearch.mean_squared_error — Method

mean_squared_error(results::AbstractVector{Tuple{<:Number,<:Number}})

Returns the mean squared error of results.

Arguments

results<:AbstractVector{<:Tuple{Number,Number}}: the vector of tuples, where each tuple is in the form Tuple{expected_output, actual_output}.

HerbSearch.misclassification — Method

misclassification(results::AbstractVector{Tuple{<:Number,<:Number}})

Returns the amount of misclassified examples, i.e. how many tuples with non-matching entries are there in results.

Arguments

results<:AbstractVector{<:Tuple{Number,Number}}: the vector of tuples, where each tuple is in the form Tuple{expected_output, actual_output}.

HerbSearch.most_likely_priority_function — Method

most_likely_priority_function(g::ContextSensitiveGrammar, tree::AbstractRuleNode, ::Union{Real, Tuple{Vararg{Real}}})

Calculates logit for all possible derivations for a node in a tree and returns them.

HerbSearch.mse_error_function — Method

mse_error_function(old_error, output, expected_output)

Mean squared error function for search_best.

- old_error         - The existing total error
- output            - The actual output of the evaluator
- expected_output   - The expected output for the example

The function build the mean squared error from output and expected_output`.

HerbSearch.mutate_random! — Function

mutate_random!(program::RuleNode, grammar::Grammar, max_depth::Int64 = 2)

Mutates the given program by inserting a randomly generated sub-program at a random location.

HerbSearch.probabilistic_accept — Method

probabilistic_accept(current_cost::Real, next_cost::Real, temperature::Real)

Probabilistically decides whether to accept the new program (next) based on the ratio of costs (smaller is better) between the previous and new program. Returns True if the new program is accepted, False otherwise.

Arguments

current_cost::Real: the cost of the current program.

next_cost::Real: the cost of the proposed program.

temperature::Real: the temperature; not used.

HerbSearch.probabilistic_accept_with_temperature — Method

probabilistic_accept_with_temperature(current_cost::Real, next_cost::Real, temperature::Real)

Returns true if the cost of the proposed program is smaller than the cost of the current program. Otherwise, returns true with the probability equal to:

\[1 / (1 + exp(delta / temperature))\]

In any other case, returns false.

Arguments

current_cost::Real: the cost of the current program.

next_cost::Real: the cost of the proposed program.

temperature::Real: the temperature of the search.

HerbSearch.probabilistic_accept_with_temperature_fraction — Method

probabilistic_accept_with_temperature_fraction(current_cost::Real, program_to_consider_cost::Real, temperature::Real)

Probabilistically decides whether to accept the new program (next) based on the ratio of costs (smaller is better) between the previous and new program multiplied by the temperature. Returns True if the new program is accepted, False otherwise.

Arguments

current_cost::Real: the cost of the current program.

next_cost::Real: the cost of the proposed program.

temperature::Real: the current temperature

HerbSearch.propagate_constraints — Function

function propagate_constraints(root::AbstractRuleNode, grammar::ContextSensitiveGrammar, local_constraints::Set{LocalConstraint}, max_holes::Int, filled_hole::Union{HoleReference, Nothing}=nothing)::Tuple{PropagateResult, Set{LocalConstraint}}

Propagates a set of local constraints recursively to all children of a given root node. As propagate_constraints gets often called when a hole was just filled, filled_hole helps keeping track to propagate the constraints to relevant nodes, e.g. children of filled_hole. max_holes makes sure that max_size of Base.iterate is not violated. The function returns the PropagateResult and the set of relevant LocalConstraints.

HerbSearch.random_fill_propose — Method

random_fill_propose(current_program, neighbourhood_node_loc, grammar, max_depth, dict)

Returns a list with only one proposed, completely random, subprogram.

Arguments

current_program::RuleNode: the current program.

neighbourhood_node_loc::NodeLoc: the location of the program to replace.

grammar::Grammar: the grammar used to create programs.

max_depth::Int: the maximum depth of the resulting programs.

dmap::AbstractVector{Int} : the minimum possible depth to reach for each rule

dict::Dict{String, Any}: the dictionary with additional arguments; not used.

HerbSearch.search — Method

search(g::Grammar, problem::Problem, start::Symbol; evaluator::Function=test_with_input, enumerator::Function=get_bfs_enumerator, max_depth::Union{Int, Nothing}=nothing, max_size::Union{Int, Nothing}=nothing, max_time::Union{Int, Nothing}=nothing, max_enumerations::Union{Int, Nothing}=nothing, allow_evaluation_errors::Bool=false)::Union{Any, Nothing}

Searches for a program by calling search_rulenode starting from Symbol start guided by enumerator and Grammar trying to satisfy the higher-order constraints in form of input/output examples defined in the Problem. This is the heart of the Herb's search for satisfying programs. Returns the found program when the evaluation calculated using evaluator is successful.

HerbSearch.search_best — Method

search_best(g::Grammar, problem::Problem, start::Symbol; evaluator::Function=test_with_input, enumerator::Function=get_bfs_enumerator, error_function::Function=default_error_function, max_depth::Union{Int, Nothing}=nothing, max_size::Union{Int, Nothing}=nothing, max_time::Union{Int, Nothing}=nothing, max_enumerations::Union{Int, Nothing}=nothing, allow_evaluation_errors::Bool=false)::Tuple{Any, Real}

Searches the grammar for the program that satisfies the maximum number of examples in the problem. The evaluator should be a function that takes a SymbolTable, expression and a dictionary with input variable assignments and returns the output of the expression.

- g                 - The grammar that defines the search space
- problem           - The problem definition with IO examples
- start             - The start symbol in the grammar
- evaluator         - The evaluation function. Takes a SymbolTable, expression and a dictionary with 
                      input variable assignments and returns the output of the expression.
- enumerator        - A constructor for the enumerator that should be used in the search
- error_function    - The error function. Takes the existing total error, the actual output of the evaluator 
                      and the expected value for the example.
- max_depth         - The maximum depth of the search
- max_time          - The maximum time allowed for the search in seconds
- max_enumerations  - The maximum number of programs to enumerate and test
- allow_evaluation_errors - Whether the search should crash if an exception is thrown in the evaluation

Returns a tuple with the best found program so far and the error. Can be considerably slower than search due to having to evaluate each expression on each example.

HerbSearch.search_rulenode — Method

search_rulenode(g::Grammar, problem::Problem, start::Symbol; evaluator::Function=test_with_input, enumerator::Function=get_bfs_enumerator, max_depth::Union{Int, Nothing}=nothing, max_size::Union{Int, Nothing}=nothing, max_time::Union{Int, Nothing}=nothing, max_enumerations::Union{Int, Nothing}=nothing, allow_evaluation_errors::Bool=false)::Union{Tuple{RuleNode, Any}, Nothing}

Searches the grammar for the program that satisfies the maximum number of examples in the problem.

    - g                 - The grammar that defines the search space
    - problem           - The problem definition with IO examples
    - start             - The start symbol in the grammar
    - evaluator         - The evaluation function. Takes a SymbolTable, expression and a dictionary with 
                          input variable assignments and returns the output of the expression.
    - enumerator        - A constructor for the enumerator that should be used in the search
    - max_depth         - The maximum depth of the search
    - max_size          - The maximum number of nodes for ASTs in the search
    - max_time          - The maximum time allowed for the search in seconds
    - max_enumerations  - The maximum number of programs to enumerate and test'
    - allow_evaluation_errors - Whether the search should crash if an exception is thrown in the evaluation
Returns a tuple of the rulenode and the expression of the solution program once it has been found, 
or nothing otherwise.

HerbSearch.select_chromosome — Method

select_chromosome(population::Array{RuleNode}, fitness_array::Array{<:Real})::RuleNode

Selects a chromosome (individual) from the population based on a fitness array. The function uses a fitness-proportionate selection strategy, often referred to as "roulette wheel" selection. Assumes fitness_array to be normalized already.

HerbSearch.select_fitness_proportional_parents — Method

select_fitness_proportional_parents(population::Array{RuleNode}, fitness_array::Array{<:Real})::Tuple{RuleNode,RuleNode}

Selects two parent chromosomes (individuals) from a population based on fitness-proportionate selection. The selected parents can be used for genetic crossover in the next steps of the algorithm.

HerbSearch.validate_iterator — Method

validate_iterator(iter)

Validates the parameters of the iterator

StatsBase.sample — Function

sample(root::RuleNode, typ::Symbol, grammar::Grammar, maxdepth::Int=typemax(Int))

Uniformly samples a random node from the tree limited to maxdepth.

StatsBase.sample — Function

sample(root::RuleNode, typ::Symbol, grammar::Grammar,
                      maxdepth::Int=typemax(Int))

Uniformly selects a random node of the given return type typ limited by maxdepth.

StatsBase.sample — Function

sample(::Type{NodeLoc}, root::RuleNode, maxdepth::Int=typemax(Int))

Uniformly selects a random node in the tree no deeper than maxdepth using reservoir sampling. Returns a NodeLoc that specifies the location using its parent so that the subtree can be replaced.

StatsBase.sample — Function

sample(::Type{NodeLoc}, root::RuleNode, typ::Symbol, grammar::Grammar)

Uniformly selects a random node in the tree of a given type, specified using its parent such that the subtree can be replaced. Returns a NodeLoc.

Index

HerbConstraints.AbstractMatchNode

HerbConstraints.ComesAfter

HerbConstraints.ComesAfter

HerbConstraints.Forbidden

HerbConstraints.ForbiddenPath

HerbConstraints.GrammarContext

HerbConstraints.LocalConstraint

HerbConstraints.LocalForbidden

HerbConstraints.LocalOneOf

HerbConstraints.LocalOrdered

HerbConstraints.MatchFail

HerbConstraints.MatchNode

HerbConstraints.MatchVar

HerbConstraints.OneOf

HerbConstraints.Ordered

HerbConstraints.PropagatorConstraint

HerbConstraints.RequireOnLeft

HerbCore.AbstractRuleNode

HerbCore.Constraint

HerbCore.Grammar

HerbCore.Hole

HerbCore.HoleReference

HerbCore.RuleNode

HerbCore.RuleNode

HerbCore.RuleNode

HerbData.IOExample

HerbData.IOPExample

HerbData.Problem

HerbGrammar.ContextFreeGrammar

HerbGrammar.ContextSensitiveGrammar

HerbGrammar.NodeLoc

HerbGrammar.SymbolTable

HerbGrammar.SymbolTable

HerbSearch.ContextSensitivePriorityEnumerator

HerbSearch.ExpandFailureReason

HerbSearch.ExpressionIterator

HerbSearch.GeneticSearchIterator

HerbSearch.PriorityQueueItem

HerbSearch.PriorityQueueItem

HerbSearch.PropagateResult

HerbSearch.StochasticSearchEnumerator

Base.get

Base.insert!

Base.isless

Base.iterate

Base.iterate

Base.iterate

Base.iterate

Base.iterate

Base.length

Base.length

Base.rand

Base.rand

Base.rand

Base.show

Base.show

HerbConstraints._matchnode2expr

HerbConstraints._matchnode2expr

HerbConstraints._matchnode2expr

HerbConstraints._matchnode2expr

HerbConstraints._pattern_match

HerbConstraints._pattern_match

HerbConstraints._pattern_match_with_hole

HerbConstraints._pattern_match_with_hole

HerbConstraints._pattern_match_with_hole

HerbConstraints._pattern_match_with_hole

HerbConstraints._rulenode_compare

HerbConstraints._rulenode_match_with_hole

HerbConstraints.addparent!

HerbConstraints.annotation2constraint

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.check_tree

HerbConstraints.copy_and_insert

HerbConstraints.make_smaller_or_equal

HerbConstraints.matchnode2expr

HerbConstraints.matchnode2expr

HerbConstraints.propagate

HerbConstraints.propagate

HerbConstraints.propagate

HerbConstraints.propagate

HerbConstraints.propagate

HerbConstraints.propagate

HerbConstraints.propagate

HerbConstraints.propagate

HerbCore.contains_hole

HerbCore.depth

HerbCore.get_node_at_location

HerbCore.get_rulesequence

HerbCore.node_depth

HerbCore.rulesoftype

HerbCore.rulesonleft

HerbCore.swap_node

HerbCore.swap_node

HerbData.read_IOPexamples

HerbData.read_IOexamples

HerbData.readdata

HerbData.readfile

HerbData.write_IOPexamples

HerbData.write_IOexamples

HerbEvaluation.evaluate_program

HerbEvaluation.execute_on_examples

HerbEvaluation.interpret

HerbEvaluation.test_all_examples

HerbEvaluation.test_examples

HerbEvaluation.test_with_input

HerbGrammar.add_rule!

HerbGrammar.addconstraint!

HerbGrammar.cfg2csg

HerbGrammar.child_types

HerbGrammar.child_types

HerbGrammar.cleanup_removed_rules!

HerbGrammar.clearconstraints!

HerbGrammar.containedin

HerbGrammar.contains_returntype

HerbGrammar.expr2cfgrammar

HerbGrammar.expr2csgrammar

HerbGrammar.expr2pcfgrammar

HerbGrammar.expr2pcsgrammar

HerbGrammar.get_childtypes

HerbGrammar.get_domain

HerbGrammar.get_domain

HerbGrammar.get_node_at_location

HerbGrammar.get_rulesequence

HerbGrammar.has_children

HerbGrammar.iscomplete

HerbGrammar.iseval

HerbGrammar.iseval

HerbGrammar.iseval

HerbGrammar.isprobabilistic

HerbGrammar.isterminal

HerbGrammar.isterminal

HerbGrammar.isterminal

HerbGrammar.isvariable

HerbGrammar.log_probability

HerbGrammar.max_arity

HerbGrammar.mindepth

HerbGrammar.mindepth_map

HerbGrammar.nchildren

HerbGrammar.nchildren

HerbGrammar.nonterminals

HerbGrammar.probability

HerbGrammar.read_cfg

HerbGrammar.read_csg

HerbGrammar.read_pcfg

HerbGrammar.read_pcsg

HerbGrammar.remove_rule!

HerbGrammar.return_type

HerbGrammar.return_type

HerbGrammar.root_node_loc

HerbGrammar.rulenode2expr

HerbGrammar.rulenode_log_probability

HerbGrammar.rulesoftype

HerbGrammar.rulesoftype

HerbGrammar.rulesoftype

HerbGrammar.rulesonleft

HerbGrammar.store_cfg

HerbGrammar.store_csg

HerbGrammar.subsequenceof

HerbGrammar.swap_node

HerbGrammar.swap_node

HerbSearch._expand

HerbSearch._expand

HerbSearch._find_next_complete_tree

HerbSearch.best_accept

HerbSearch.calculate_cost

HerbSearch.const_temperature

HerbSearch.constructNeighbourhood

HerbSearch.constructNeighbourhoodRuleSubset

HerbSearch.count_expressions

HerbSearch.count_expressions

HerbSearch.crossover_swap_children_1

HerbSearch.crossover_swap_children_2

HerbSearch.decreasing_temperature

HerbSearch.default_error_function

HerbSearch.default_fitness

HerbSearch.enumerate_neighbours_propose

HerbSearch.get_best_program

HerbSearch.get_bfs_enumerator

HerbSearch.get_bfs_enumerator

HerbSearch.get_dfs_enumerator

HerbSearch.get_dfs_enumerator

HerbSearch.get_genetic_enumerator

HerbSearch.get_mh_enumerator

HerbSearch.get_most_likely_first_enumerator

HerbSearch.get_most_likely_first_enumerator

HerbSearch.get_sa_enumerator

HerbSearch.get_vlsn_enumerator

HerbSearch.heuristic_leftmost

HerbSearch.heuristic_random

HerbSearch.heuristic_rightmost

HerbSearch.heuristic_smallest_domain

HerbSearch.mean_squared_error

HerbSearch.misclassification

HerbSearch.most_likely_priority_function

HerbSearch.mse_error_function

HerbSearch.mutate_random!

HerbSearch.probabilistic_accept

HerbSearch.probabilistic_accept_with_temperature

HerbSearch.probabilistic_accept_with_temperature_fraction

HerbSearch.propagate_constraints

HerbSearch.random_fill_propose

HerbSearch.search

HerbSearch.search_best

HerbSearch.search_rulenode

HerbSearch.select_chromosome

HerbSearch.select_fitness_proportional_parents

HerbSearch.validate_iterator

StatsBase.sample

StatsBase.sample

StatsBase.sample

StatsBase.sample

HerbConstraints.@csgrammar_annotated

HerbGrammar.@cfgrammar

HerbGrammar.@csgrammar

HerbGrammar.@pcfgrammar

HerbGrammar.@pcsgrammar